% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_model.R
\name{add_strategies}
\alias{add_strategies}
\alias{add_mutant}
\alias{expand_parameters}
\alias{mutant_parameters}
\title{Add strategies (or a mutant) to Parameters}
\usage{
add_strategies(
  p,
  traits,
  hyperpar = param_hyperpar(p),
  birth_rate = 1,
  keep_existing = TRUE
)

add_mutant(p, traits, hyperpar = param_hyperpar(p), birth_rate = 1)

expand_parameters(
  trait_matrix,
  p,
  hyperpar = param_hyperpar(p),
  birth_rate_list = 1,
  keep_existing_strategies = TRUE
)

mutant_parameters(
  trait_matrix,
  p,
  hyperpar = param_hyperpar(p),
  birth_rate_list = 1,
  keep_existing_strategies = FALSE
)
}
\arguments{
\item{p}{A \code{Parameters} object.}

\item{traits}{A matrix of traits corresponding to the new strategies to
introduce (see \code{\link{trait_matrix}}).}

\item{hyperpar}{Hyperparameter function to use. By default links to the
standard function for this strategy type.}

\item{birth_rate}{Birth rate(s), one per row of \code{traits}. See
\code{\link{generate_strategy}}.}

\item{keep_existing}{Should existing resident strategies be retained?}
}
\description{
Add strategies to a \code{Parameters} object. \code{add_strategies} appends
the new strategies to any existing residents (controlled by
\code{keep_existing}); \code{add_mutant} introduces strategies as mutants
(i.e. replacing the resident set). Both translate trait values into
strategies via \code{\link{generate_strategy}} and are pipe-friendly
(the \code{Parameters} object is the first argument).
}
