% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assembler_helper}
\alias{assembler_helper}
\title{Assembler Helper}
\usage{
assembler_helper(sys0, births = "naive", ..., vcv_p = 0.001)
}
\arguments{
\item{sys0}{Initial state of the system}

\item{births}{Strategy for births (naive or sampling)}

\item{...}{Additional arguments passed to either
\code{\link{assembler_stochastic_naive}} or to
\code{\link{assembler_sample_positive}}.  In particular,
\code{compute_viable_fitness}, \code{run_type},
\code{jump_to_attractor} are supported by both and
\code{n_mutants}, \code{n_immigrants} and \code{n_sample} are
supported by the underlying birth functions.}

\item{vcv_p}{Mutantional variance as a fraction of the landscape
bounds}
}
\description{
Wrapper around assembler that tries to make things a bit easier.
}
\details{
If the strategy is naive, then
}
\author{
Rich FitzJohn
}

