% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scm_support.R
\name{assembly_parameters}
\alias{assembly_parameters}
\title{Helper function for creating parameter objects}
\usage{
assembly_parameters(..., pars = NULL)
}
\arguments{
\item{...}{Named set of parameters}

\item{pars}{A list of parameters}
}
\description{
Helper function for creating parameter objects suitable for an
assembly.
}

