% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bounds}
\alias{bounds}
\title{Trait bounds}
\usage{
bounds(...)
}
\arguments{
\item{...}{Named list, each element of which is a 2-element
numeric vector of lower and upper bounds.}
}
\description{
Helper function for making bounds
}
\examples{
bounds(lma=c(0.01, 10))
bounds(lma=c(0.01, 10), rho=c(1, 1000))
}

