% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build_schedule_seed_rain}
\alias{build_schedule_seed_rain}
\title{Build Schedule From Seed Rain Changes}
\usage{
build_schedule_seed_rain(p, schedule = NULL)
}
\arguments{
\item{p}{Parameters object}

\item{times}{Vector of times to start from}

\item{nsteps}{Number of new cohorts to add}

\item{progress}{Save progress with the returned times?}

\item{verbose}{Print information about progress as we go?}
}
\value{
Numeric vector of times
}
\description{
Build a schedule of cohort times by iteratively finding the
position that causes the greatest change in seed_rain.  In contrast
with \code{\link{build_schedule}}, this is not really designed for
use in real analyses, but instead to confirm that the schedules
work as expected.  It is \emph{extremely slow} because every step
requires introducing a cohort between every pair of cohorts.
}
\author{
Rich FitzJohn
}

