% Generated by roxygen2 (4.0.2): do not edit by hand
\name{carrying_capacity}
\alias{carrying_capacity}
\title{Carrying Capacity}
\usage{
carrying_capacity(trait, values, p, seed_rain = 1, parallel = FALSE)
}
\arguments{
\item{trait}{Name of the trait (e.g., \code{"lma"})}

\item{values}{Values to compute maximum growth rate for}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{seed_rain}{Initial seed rain (optional)}

\item{parallel}{Use multiple processors?}
}
\description{
Compute the carrying capacity (equilibrium per-capita seed
production) for a set of values of a trait.  Each is considered in
isolation.
}
\author{
Rich FitzJohn
}

