% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{check_inviable}
\alias{check_inviable}
\title{Check low-abundance strategies for viability}
\usage{
check_inviable(p, ctrl)
}
\arguments{
\item{p}{A Parameters object}

\item{ctrl}{Control object}
}
\description{
Check low-abundance strategies for viability.
}
