% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/util.R
\name{clamp_domain}
\alias{clamp_domain}
\title{Clamp function to domain}
\usage{
clamp_domain(f, r, value = NA_real_)
}
\arguments{
\item{f}{A function that takes \code{x} as a first argument.}

\item{r}{Range of values (vector of length 2)}

\item{value}{(Single) value to use when out of domain.}
}
\value{
A new function
}
\description{
Clamp a function to particular values when outside of a given domain (r)
}
\details{
Things like names on input and return vectors are not dealt with
very well, and would differ if the function was better
constructed.  Values falling outwide the domain are not evaluated
(which is useful if these would cause crashes, warnings, etc).
}
