% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{CohortSchedule}
\alias{CohortSchedule}
\title{Schedule of cohort introduction times}
\usage{
CohortSchedule(n_species)
}
\arguments{
\item{n_species}{number of species}
}
\description{
Schedule of cohort introduction times
}
