% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Control}
\alias{Control}
\title{Control parameters}
\usage{
Control(..., values = list(...))
}
\arguments{
\item{...,values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Control parameters that tune various aspects of the numerical solvers.
}

