% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default_cohort_schedule}
\alias{default_cohort_schedule}
\title{Generate Default Cohort Schedule}
\usage{
default_cohort_schedule(p)
}
\arguments{
\item{p}{A Parameters object}
}
\description{
Generate default cohort schedule.  This has times compressed
towards young patch ages, which usually need finer resolution than
later patche ages.  Currently there is no support for arbitrary
times: just build the schedule manually for now.
}
\author{
Rich FitzJohn
}

