% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Disturbance}
\alias{Disturbance}
\title{Disturbance control object}
\usage{
Disturbance(mean_interval)
}
\arguments{
\item{mean_interval}{Mean disturbance interval in years}
}
\description{
Disturbance control object
}
