% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ebt_base_parameters}
\alias{ebt_base_parameters}
\title{Sensible, fast (ish) EBT parameters}
\usage{
ebt_base_parameters()
}
\description{
Hopefully sensible set of parameters for use with the EBT.  Turns
accuracy down a bunch, makes it noisy.
}
\author{
Rich FitzJohn
}

