% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_support.R
\name{environment_type}
\alias{environment_type}
\alias{Environment}
\title{Creates an environment object of specified type}
\usage{
environment_type(type)

Environment(type = NULL)
}
\arguments{
\item{type}{Any environment name as a string, e.g.: \code{"FF16_Env"}.}
}
\description{
Creates an environment object of specified type
}
