% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{equilibrium_birth_rate}
\alias{equilibrium_birth_rate}
\title{Run system to offspring arrival equilibrium}
\usage{
equilibrium_birth_rate(p, ctrl)
}
\arguments{
\item{p}{A \code{Parameters} object}

\item{ctrl}{Control object}
}
\value{
A Parameters object, with offspring arrival and node schedule
elements set.
}
\description{
Run system to offspring arrival equilibrium
}
\author{
Rich FitzJohn
}
