% Generated by roxygen2 (4.0.2): do not edit by hand
\name{equilibrium_seed_rain}
\alias{equilibrium_seed_rain}
\title{Determine Equilibrium Seed Rain}
\usage{
equilibrium_seed_rain(p, schedule = NULL)
}
\arguments{
\item{p}{Parameters object}

\item{schedule}{Schedule to run EBT with, or \code{NULL} for a
default schedule (should generally be a reasonable choice as it's
only the starting seed for the usual schedule building with
\code{\link{build_schedule}}).}
}
\description{
Determine the equilibrium seed rain for a set of strategies.
}
\details{
This does not use any clever root finding algorithm to try and
speed up convergence, but simply iterates the solver.  Something
like the secant method would speed this up greatly and help in
divergent cases.
}
\author{
Rich FitzJohn
}

