% Generated by roxygen2 (4.0.0): do not edit by hand
\name{equilibrium_verbose}
\alias{equilibrium_verbose}
\title{Noisy Parameters for Equilibrium Finding}
\usage{
equilibrium_verbose()
}
\description{
Control parameters for \code{\link{equilibrium_seed_rain}} that
make progress noisier.  This is just a convenience function.
}
\examples{
p <- new(Parameters)
p$set_control_parameters(equilibrium_verbose())
}

