% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/es20.R
\name{ES20_Plant}
\alias{ES20_Plant}
\alias{ES20_Cohort}
\alias{ES20_Species}
\alias{ES20_Parameters}
\alias{ES20_Patch}
\alias{ES20_SCM}
\alias{ES20_StochasticSpecies}
\alias{ES20_StochasticPatch}
\alias{ES20_StochasticPatchRunner}
\alias{ES20_PlantPlus}
\title{Create a ES20 Plant or Cohort}
\usage{
ES20_Plant(s = ES20_Strategy())

ES20_Cohort(s = ES20_Strategy())

ES20_Species(s = ES20_Strategy())

ES20_Parameters(...)

ES20_Patch(p)

ES20_SCM(p)

ES20_StochasticSpecies(s = ES20_Strategy())

ES20_StochasticPatch(p)

ES20_StochasticPatchRunner(p)

ES20_PlantPlus(s = ES20_Strategy())
}
\arguments{
\item{s}{A \code{\link{ES20_Strategy}} object}

\item{...}{Arguments!}

\item{p}{A \code{Parameters<ES20>} object}
}
\description{
Create a ES20 Plant or Cohort
}
\examples{
pl <- ES20_Plant()
pl$height
}
