% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es20r.R
\name{ES20r_Plant}
\alias{ES20r_Plant}
\alias{lcp_whole_plant.Plant<ES20r>}
\alias{ES20r_Cohort}
\alias{ES20r_Species}
\alias{ES20r_Parameters}
\alias{ES20r_Patch}
\alias{ES20r_SCM}
\alias{ES20r_StochasticSpecies}
\alias{ES20r_StochasticPatch}
\alias{ES20r_StochasticPatchRunner}
\title{Create a ES20r Plant or Cohort}
\usage{
ES20r_Plant(s = ES20r_Strategy())

\method{lcp_whole_plant}{`Plant<ES20r>`}(p, ...)

ES20r_Cohort(s = ES20r_Strategy())

ES20r_Species(s = ES20r_Strategy())

ES20r_Parameters()

ES20r_Patch(p)

ES20r_SCM(p)

ES20r_StochasticSpecies(s = ES20r_Strategy())

ES20r_StochasticPatch(p)

ES20r_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{ES20r_Strategy}} object}

\item{p}{A \code{Parameters<ES20r,ES20r_Env>} object}

\item{...}{Arguments!}
}
\description{
Create a ES20r Plant or Cohort
}
\examples{
pl <- ES20r_Plant()
pl$height
}
