% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/es20r.R
\name{ES20r_Environment}
\alias{ES20r_Environment}
\alias{ES20r_make_environment}
\title{FF16r_Environment object}
\usage{
ES20r_Environment(disturbance_mean_interval, seed_rain, k_I, control)

ES20r_make_environment(p)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{k_I}{light extinction coefficient}

\item{control}{A Control object}

\item{p}{A Parameters object}
}
\description{
FF16r_Environment object
}
