% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R, R/strategy_support.R, R/tf24.R
\name{FF16_expand_state}
\alias{FF16_expand_state}
\alias{expand_state}
\alias{TF24_expand_state}
\title{Add additional state variables to the species component in output of a model.}
\usage{
FF16_expand_state(tidy_patch_results)

expand_state(results_tidy)

TF24_expand_state(tidy_patch_results)
}
\arguments{
\item{tidy_patch_results}{from `tidy_patch`}
}
\value{
similar format to input, but with additional columns for additional state variables
}
\description{
Add additional state variables to the species component in output of a model.
}
