% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_expand_state.R
\name{expand_state_FF16}
\alias{expand_state_FF16}
\title{Add additonal state variables to the species component in output of FF16 model.}
\usage{
expand_state_FF16(tidy_patch_results)
}
\arguments{
\item{output}{from `tidy_patch`}
}
\value{
similar format to input, but with additional columns for additonal state variables
}
\description{
Add additonal state variables to the species component in output of FF16 model.
}
