% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fast_control}
\alias{fast_control}
\title{Fast Control Defaults}
\usage{
fast_control(base = Control())
}
\arguments{
\item{base}{An optional \code{Control} object.  If omitted, the
defaults are used.}
}
\value{
A Control object with parameters set.
}
\description{
Fast Control Defaults
}
\author{
Rich FitzJohn
}

