% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R
\name{FF16_Individual}
\alias{FF16_Individual}
\alias{lcp_whole_plant.Plant<FF16>}
\alias{FF16_Cohort}
\alias{FF16_Species}
\alias{FF16_Parameters}
\alias{FF16_Patch}
\alias{FF16_SCM}
\alias{FF16_StochasticSpecies}
\alias{FF16_StochasticPatch}
\alias{FF16_StochasticPatchRunner}
\title{Create a FF16 Plant or Cohort}
\usage{
FF16_Individual(s = FF16_Strategy())

\method{lcp_whole_plant}{`Plant<FF16>`}(p, ...)

FF16_Cohort(s = FF16_Strategy())

FF16_Species(s = FF16_Strategy())

FF16_Parameters()

FF16_Patch(p)

FF16_SCM(p)

FF16_StochasticSpecies(s = FF16_Strategy())

FF16_StochasticPatch(p)

FF16_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{FF16_Strategy}} object}

\item{p}{A \code{Parameters<FF16,FF16_Env>} object}

\item{...}{Arguments!}
}
\description{
Create a FF16 Plant or Cohort
}
\examples{
pl <- FF16_Individual()
pl$height
}
