% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_expand_state.R
\name{FF16_expand_state}
\alias{FF16_expand_state}
\title{Add additonal state variables to the species component in output of FF16 model.}
\usage{
FF16_expand_state(tidy_patch_results)
}
\arguments{
\item{output}{from `tidy_patch`}
}
\value{
similar format to input, but with additional columns for additonal state variables
}
\description{
Add additonal state variables to the species component in output of FF16 model.
}
