% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R, r/ff16w.R, r/k93.R, r/tf24.R
\name{FF16_fixed_environment}
\alias{FF16_fixed_environment}
\alias{FF16w_fixed_environment}
\alias{K93_fixed_environment}
\alias{TF24_fixed_environment}
\title{Construct a fixed environment for a model}
\usage{
FF16_fixed_environment(e = 1, height_max = 150, ...)

FF16w_fixed_environment(e = 1, height_max = 150)

K93_fixed_environment(e = 1, height_max = 300, ...)

TF24_fixed_environment(e = 1, height_max = 150, ...)
}
\arguments{
\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}

\item{...}{Additional parameters to be passed to \code{XXXX_make_environment}, where XXXX referes to model name.}
}
\description{
Construct a fixed environment for a model

Construct a fixed environment for FF16w strategy
}
