% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R, R/ff16r.R, R/k93.R
\name{FF16_make_environment}
\alias{FF16_make_environment}
\alias{FF16r_make_environment}
\alias{K93_make_environment}
\title{: Helper functions to create an FF16_Environment object. Useful for running individuals}
\usage{
FF16_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)

FF16r_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)

K93_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)
}
\arguments{
\item{light_availability_spline_tol}{Error tolerance of adpative spline method. Deafult is 1e-4.}

\item{light_availability_spline_nbase}{Parameter used in adaptive spline method. Default is 17.}

\item{light_availability_spline_max_depth}{Parameter used in adaptive spline method. Default is 16.}

\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}
}
\description{
: Helper functions to create an FF16_Environment object. Useful for running individuals
}
