% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{FF16_Pars}
\alias{FF16_Pars}
\title{FF16 strategy parameters}
\usage{
FF16_Pars(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Biological trait/parameters for the FF16 strategy. Held as the nested
\code{pars} element of an \code{\link{FF16_Strategy}} (access as
\code{s$pars$lma}).
}
