% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16bg.R
\name{FF16bg_Individual}
\alias{FF16bg_Individual}
\alias{lcp_whole_plant.Plant<FF16bg>}
\alias{FF16bg_Cohort}
\alias{FF16bg_Species}
\alias{FF16bg_Parameters}
\alias{FF16bg_Patch}
\alias{FF16bg_SCM}
\alias{FF16bg_StochasticSpecies}
\alias{FF16bg_StochasticPatch}
\alias{FF16bg_StochasticPatchRunner}
\title{Create a FF16bg Plant or Cohort}
\usage{
FF16bg_Individual(s = FF16bg_Strategy())

`lcp_whole_plant.Plant<FF16bg>`(p, ...)

FF16bg_Cohort(s = FF16bg_Strategy())

FF16bg_Species(s = FF16bg_Strategy())

FF16bg_Parameters()

FF16bg_Patch(p)

FF16bg_SCM(p)

FF16bg_StochasticSpecies(s = FF16bg_Strategy())

FF16bg_StochasticPatch(p)

FF16bg_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{FF16bg_Strategy}} object}

\item{p}{A \code{Parameters<FF16bg,FF16_Env>} object}

\item{...}{Arguments!}
}
\description{
Create a FF16bg Plant or Cohort
}
\examples{
pl <- FF16bg_Individual()
pl$height
}
