% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16r.R
\name{FF16r_test_environment}
\alias{FF16r_test_environment}
\title{Create a test environment for FF16r startegy}
\usage{
FF16r_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)
}
\arguments{
\item{height}{top height of environment object}

\item{n}{number of points}

\item{light_env}{function for light environment in test object}

\item{n_strategies}{number of strategies for test environment}
}
\description{
This makes a pretend light environment over the plant height,
slightly concave up, whatever.
}
\examples{
environment <- FF16r_test_environment(10)
}
