% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16w.R
\name{FF16w_Individual}
\alias{FF16w_Individual}
\alias{FF16w_Cohort}
\alias{FF16w_Species}
\alias{FF16w_Parameters}
\alias{FF16w_Patch}
\alias{FF16w_SCM}
\alias{FF16w_StochasticSpecies}
\alias{FF16w_StochasticPatch}
\alias{FF16w_StochasticPatchRunner}
\title{Create a FF16w Plant or Cohort}
\usage{
FF16w_Individual(s = FF16w_Strategy())

FF16w_Cohort(s = FF16w_Strategy())

FF16w_Species(s = FF16w_Strategy())

FF16w_Parameters()

FF16w_Patch(p)

FF16w_SCM(p)

FF16w_StochasticSpecies(s = FF16w_Strategy())

FF16w_StochasticPatch(p)

FF16w_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{FF16w_Strategy}} object}

\item{p}{A \code{Parameters<FF16w,FF16_Env>} object}
}
\description{
Create a FF16w Plant or Cohort
}
\examples{
pl <- FF16w_Individual()
pl$height
}
