% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16w.R
\name{make_FF16w_hyperpar}
\alias{make_FF16w_hyperpar}
\alias{FF16w_hyperpar}
\title{Hyperparameters for FF16w physiological model}
\usage{
make_FF16w_hyperpar(
  lma_0 = 0.1978791,
  B_kl1 = 0.4565855,
  B_kl2 = 1.71,
  rho_0 = 608,
  B_dI1 = 0.01,
  B_dI2 = 0,
  B_ks1 = 0.2,
  B_hks1 = 0.2,
  B_hks2 = 0,
  B_ks2 = 0,
  B_rs1 = 4012,
  B_rb1 = 2 * 4012,
  B_f1 = 3,
  a_lf1 = 0.535,
  B_lf1 = 0.009,
  B_lf2 = 0.004,
  B_lf3 = 8e-04,
  B_lf4 = 21000,
  B_lf5 = 40000,
  k_I = 0.5,
  latitude = 0,
  B_Hv1 = 1.731347,
  B_Hv2 = -0.7246377,
  B_c1 = 2.04,
  B_c2 = 0,
  K_s_0 = 2
)

FF16w_hyperpar(m, s, filter = TRUE)
}
\arguments{
\item{lma_0}{Central (mean) value for leaf mass per area [kg /m2]}

\item{B_kl1}{Rate of leaf turnover at lma_0 [/yr]}

\item{B_kl2}{Scaling slope for phi in leaf turnover [dimensionless]}

\item{rho_0}{Central (mean) value for wood density [kg /m3]}

\item{B_dI1}{Rate of instantaneous mortality at rho_0 [/yr]}

\item{B_dI2}{Scaling slope for wood density in intrinsic mortality [dimensionless]}

\item{B_ks1}{Rate of sapwood turnover at rho_0 [/yr]}

\item{B_ks2}{Scaling slope for rho in sapwood turnover [dimensionless]}

\item{B_rs1}{CO_2 respiration per unit sapwood volume [mol / yr / m3 ]}

\item{B_rb1}{CO_2 respiration per unit sapwood volume [mol / yr / m3 ]}

\item{B_f1}{Cost of seed accessories per unit seed mass [dimensionless]}

\item{a_lf1}{intercept for empirical relationship between narea and vcmax, lma (Dong et al. 2022)}

\item{B_lf1}{Beta coefficient for empirical relationship between narea ~ lma [g/m2] (Dong et al. 2022)}

\item{B_lf2}{Beta coefficient for empirical relationship between narea ~ vcmax [umol / m2 / s] (Dong et al. 2022)}

\item{B_lf4}{CO_2 respiration per unit leaf nitrogen [mol / yr / kg]}

\item{k_I}{light extinction coefficient [dimensionless]}

\item{latitude}{degrees from equator (0-90), used in solar model [deg]}

\item{B_Hv1}{p50 at K_s_0 [-MPa]}

\item{B_Hv2}{Scaling slope for K_s in p50 [dimensionless]}

\item{K_s_0}{Central (mean) value for maximum sapwood conductivity [kg /m2 / s / MPA]}

\item{m}{A matrix of trait values, as returned by \code{trait_matrix}}

\item{s}{A strategy object}

\item{filter}{A flag indicating whether to filter columns. If TRUE, any numbers
that are within eps of the default strategy are not replaced.}
}
\description{
Hyperparameters for FF16w physiological model

Hyperparameter function for FF16w physiological model
}
