% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffdev.R
\name{FFdev_Plant}
\alias{FFdev_Cohort}
\alias{FFdev_EBT}
\alias{FFdev_Parameters}
\alias{FFdev_Patch}
\alias{FFdev_Plant}
\alias{FFdev_PlantPlus}
\alias{FFdev_Species}
\alias{FFdev_StochasticPatch}
\alias{FFdev_StochasticPatchRunner}
\alias{FFdev_StochasticSpecies}
\title{Create a FFdev Plant or Cohort}
\usage{
FFdev_Plant(s = FFdev_Strategy())

FFdev_Cohort(s = FFdev_Strategy())

FFdev_Species(s = FFdev_Strategy())

FFdev_Parameters(...)

FFdev_Patch(p)

FFdev_EBT(p)

FFdev_StochasticSpecies(s = FFdev_Strategy())

FFdev_StochasticPatch(p)

FFdev_StochasticPatchRunner(p)

FFdev_PlantPlus(s = FFdev_Strategy())
}
\arguments{
\item{s}{A \code{\link{FFdev_Strategy}} object}

\item{...}{Arguments!}

\item{p}{A \code{Parameters<FFdev>} object}
}
\description{
Create a FFdev Plant or Cohort
}
\examples{
pl <- FFdev_Plant()
pl$height
}

