% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffw16.R
\name{FFW16_Plant}
\alias{FFW16_Cohort}
\alias{FFW16_Parameters}
\alias{FFW16_Plant}
\alias{FFW16_Species}
\title{Create a FFW16 Plant or Cohort}
\usage{
FFW16_Plant(s = FFW16_Strategy())

FFW16_Cohort(s = FFW16_Strategy())

FFW16_Species(s = FFW16_Strategy())

FFW16_Parameters(...)
}
\arguments{
\item{s}{A \code{\link{FFW16_Strategy}} object}

\item{...}{Arguments!}
}
\description{
Create a FFW16 Plant or Cohort
}
\examples{
pl <- FFW16_Plant()
pl$height
}

