% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{FFW16_Parameters}
\alias{FFW16_Parameters}
\title{Strategy parameters}
\usage{
FFW16_Parameters(..., values = list(...))
}
\arguments{
\item{...,values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Strategy parameters that tune various aspects of the biological model.
}

