% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffw16.R
\name{FFW16_Plant}
\alias{FFW16_Plant}
\title{Create a FFW16 Plant}
\usage{
FFW16_Plant(s = FFW16_Strategy())
}
\arguments{
\item{s}{A \code{\link{FFW16_Strategy}} object}
}
\description{
Create a FFW16 Plant
}
\examples{
pl <- FFW16_Plant()
pl$height
}

