% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_singularity_1D}
\alias{find_singularity_1D}
\title{Find evolutionary attractor for single species and trait.}
\usage{
find_singularity_1D(trait, interval, p, tol = 1e-04, ...)
}
\arguments{
\item{trait}{name of trait}

\item{interval}{a vector containing the end-points of the
interval to be searched for the root.}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{tol=1e-4}{the desired accuracy (convergence tolerance).}

\item{...}{set verbpse=TRUE for verbose output, seed_rain=value
gives starting values when solving for demographic equilibrium}
}
\value{
Same as \code{uniroot} function: A list with at least
four components:
‘root’ and ‘f.root’ give the location of the root and the
value of the function evaluated at that point. ‘iter’ and
‘estim.prec’ give the number of iterations
}
\description{
Find evolutionary attractor for single species and trait.
This is point at which selection gradient equals zero.
Currently solved using \code{uniroot}
}
\author{
Daniel Falster
}

