% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/fitness.R
\name{fitness_landscape}
\alias{fitness_landscape}
\title{Fitness Landscape}
\usage{
fitness_landscape(
  trait_matrix,
  p,
  hyperpar = param_hyperpar(p),
  raw_seed_rain = FALSE
)
}
\arguments{
\item{trait_matrix}{A matrix of traits corresponding to mutants to
introduce into the light environment constructed by the residents
in \code{p}.}

\item{p}{Parameters object.  Needs to contain residents with their
incoming seed rain.}

\item{raw_seed_rain}{Logical; if \code{TRUE} report per capita
seed rain rather than fitness.}
}
\value{
Vector with the output seed rain.  Mutants have an
arbitrary seed rain of one, so this is the rate of seed
production per capita.
}
\description{
Construct a fitness landscape.
}
\author{
Rich FitzJohn
}
