% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixed_environment}
\alias{fixed_environment}
\title{Create fixed light environment}
\usage{
fixed_environment(canopy_openness, height_max = 150)
}
\arguments{
\item{canopy_openness}{Index of canopy openness (on 0,1)}

\item{height_max}{Maximum height.  The default (150) should be big
enough for most uses.}
}
\description{
Create a light environment where light levels are constant down
the canopy.
}
\author{
Rich FitzJohn
}
