% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_equilibrium_community}
\alias{get_equilibrium_community}
\title{Solve for equilbrium community with given values of specified trait}
\usage{
get_equilibrium_community(trait, values, p, seed_rain = NULL)
}
\arguments{
\item{trait}{name of trait}

\item{values}{vector of trait values for community}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{seed_rain=NULL}{vector of initial seed rains for community}

\item{verbose=TRUE}{prints details to screen}
}
\description{
Find seed rain and cohort schedule for equilbrium community with
given traits. This is point at which each resident seed returns
on average a single seed.
}
\author{
Daniel Falster
}

