% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gradient_fd}
\alias{gradient_fd}
\title{Calculate the gradient of a function by numerical approximation}
\usage{
gradient_fd(func, x, dx, log_scale = FALSE)
}
\arguments{
\item{func}{a function with a scalar real result}

\item{x}{a real scalar or vector argument to func, indicating
the point(s) at which the gradient is to be calculated.}

\item{dx}{Interval over which derivative is calculated}

\item{log_scale=TRUE}{Determines whether derivative is taken
with respect to raw or log-transformed x values. The latter is
equiavlent to taking the derivative of a function
\code{g(f(exp(x)))} with respect to log x and is useful when
x is log-normally distributed.}
}
\description{
The function ‘gradient_fd’ calculates a numerical approximation
of the first derivative of ‘func’ at the point ‘x’ using finite
differencing. It is assumed ‘func’ is a scalar value function.
}
\author{
Daniel Falster
}

