% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grow.plant.to.size}
\alias{grow.plant.to.size}
\title{Grow Plant To Given Size}
\usage{
grow.plant.to.size(plant, env, distance.from.target, max.time = 1000)
}
\arguments{
\item{plant}{A \code{Plant} object.}

\item{env}{An \code{Environment} object.}

\item{distance.from.target}{A function that takes \code{plant} as
its only argument and computes the distance from a target.  The
function must return negative values when the plant is smaller
than the target.  See Details.}

\item{max.time}{Time to run the ODE out for -- only exists to
prevent an infinite loop (say, on an unreachable size).}
}
\value{
A plant grown to the appropriate size.  This is a
\emph{copy} of the input plant.
}
\description{
Grow a plant up to a certain size from a seed.
}
\details{
This is a little utility and is likely to fail.

To stop at a height of 10, the \code{distance.from.target}
function should be \code{function(plant) plant$height - 10}, so
that the return value is negative when the plant is smaller than
the target size.
}
\author{
Rich FitzJohn
}

