% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\name{grow_plant_to_time_expanded}
\alias{grow_plant_to_time_expanded}
\title{Grow a plant}
\usage{
grow_plant_to_time_expanded(plant, times, env)
}
\arguments{
\item{plant}{A \code{Plant} object}

\item{times}{A vector of times}

\item{env}{An \code{Environment} object}
}
\description{
Grow a plant up for particular time lengths
Save auxilary data as well as state data
}
