% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{Individual}
\alias{Individual}
\title{Individual object}
\usage{
Individual(T, E)
}
\arguments{
\item{T}{String containing class of plant strategy to create;}

\item{E}{String containing class of environmnet to create;}
}
\description{
Individual object
}
