% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_outputs.R
\name{integrate_over_size_distribution}
\alias{integrate_over_size_distribution}
\title{Integrates over size distribution for each species at each time point, to give totals of each variable
Integrations are performed using trapezium integration}
\usage{
integrate_over_size_distribution(tidy_species_data)
}
\arguments{
\item{tidy_species_data}{output of either `tidy_patch` or `tidy_species`}
}
\value{
a tibble whose columns provide metrics on integrated totals for each variable for each species at each time
}
\description{
Integrates over size distribution for each species at each time point, to give totals of each variable
Integrations are performed using trapezium integration
}
