% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Internals}
\alias{Internals}
\title{Extract Internals from plant object}
\usage{
Internals(s_size, a_size)
}
\arguments{
\item{s_size}{???}

\item{a_size}{???}
}
\description{
Internals class holding vectors of states, thier associated rates and auxiliary (aux) state
which is calculated from the state through running compute_rates
}
