% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_patch.R
\name{interpolate_to_times}
\alias{interpolate_to_times}
\title{Interpolate tidyspecies data to specific time points}
\usage{
interpolate_to_times(tidy_species_data, times, method = "natural")
}
\arguments{
\item{tidy_species_data}{}

\item{times}{times to interpolate to}

\item{method}{Method for interpolation. For more info see help on stats::spline}
}
\value{
??
}
\description{
Interpolate tidyspecies data to specific time points
}
