% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/k93.R
\name{K93_Plant}
\alias{K93_Plant}
\alias{lcp_whole_plant.Plant<K93>}
\alias{K93_Cohort}
\alias{K93_Species}
\alias{K93_Parameters}
\alias{K93_Patch}
\alias{K93_SCM}
\alias{K93_StochasticSpecies}
\alias{K93_StochasticPatch}
\alias{K93_StochasticPatchRunner}
\title{Create a K93 Plant or Cohort}
\usage{
K93_Plant(s = K93_Strategy())

\method{lcp_whole_plant}{`Plant<K93>`}(p, ...)

K93_Cohort(s = K93_Strategy())

K93_Species(s = K93_Strategy())

K93_Parameters()

K93_Patch(p)

K93_SCM(p)

K93_StochasticSpecies(s = K93_Strategy())

K93_StochasticPatch(p)

K93_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{K93_Strategy}} object}

\item{p}{A \code{Parameters<K93,K93_Env>} object}

\item{...}{Arguments!}
}
\description{
Create a K93 Plant or Cohort
}
\examples{
pl <- K93_Plant()
pl$height
}
