% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R, r/k93.R
\name{K93_Environment}
\alias{K93_Environment}
\alias{K93_make_environment}
\alias{K93_fixed_environment}
\title{K93_Environment object}
\usage{
K93_Environment(control)

K93_make_environment(ctrl = scm_base_control())

K93_fixed_environment(e = 1, ctrl = scm_base_control(), height_max = 300)
}
\arguments{
\item{control}{A Control object}

\item{ctrl}{Control object}

\item{e}{Value of environment (default=1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}
}
\description{
K93_Environment object

Construct a fixed environment for K93 strategy
}
