% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k93.R
\name{K93_hyperpar}
\alias{K93_hyperpar}
\title{Hyperparameter function for K93 physiological model}
\usage{
K93_hyperpar(m, s, filter = TRUE)
}
\arguments{
\item{m}{A matrix of trait values, as returned by \code{trait_matrix}}

\item{s}{A strategy object}

\item{filter}{A flag indicating whether to filter columns. If TRUE, any numbers
that are within eps of the default strategy are not replaced.}
}
\description{
Hyperparameter function for K93 physiological model
}
