% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/k93.R
\name{K93_Individual}
\alias{K93_Individual}
\title{Create a K93 Individual or Node}
\usage{
K93_Individual(s = K93_Strategy())
}
\arguments{
\item{s}{A \code{\link{K93_Strategy}} object}
}
\description{
Create a K93 Individual or Node
}
\examples{
pl <- K93_Individual()
pl$height
}
