% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k93.R
\name{K93_make_environment}
\alias{K93_make_environment}
\title{create K93_environment object}
\usage{
K93_make_environment(
  canopy_light_tol = 1e-04,
  canopy_light_nbase = 17,
  canopy_light_max_depth = 16,
  canopy_rescale_usually = TRUE
)
}
\arguments{
\item{canopy_light_tol}{}

\item{canopy_rescale_usually}{}
}
\description{
create K93_environment object
}
