% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{K93_Pars}
\alias{K93_Pars}
\title{K93 strategy parameters}
\usage{
K93_Pars(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Biological trait/parameters for the K93 strategy. Held as the nested
\code{pars} element of a \code{\link{K93_Strategy}} (access as
\code{s$pars$b_0}).
}
