% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k93.R
\name{K93_test_environment}
\alias{K93_test_environment}
\title{Create a test environment for K93 startegy}
\usage{
K93_test_environment(
  height,
  n = 101,
  light_env = NULL,
  n_strategies = 1,
  seed_rain = 0
)
}
\arguments{
\item{height}{top height of environment object}

\item{n}{number of points}

\item{light_env}{function for light environment in test object}

\item{n_strategies}{number of strategies for test environment}

\item{seed_rain}{seed_rain for test environment}
}
\description{
This makes a pretend light environment over the plant height,
slightly concave up, whatever.
}
\examples{
environment <- K93_test_environment(10)
}
