% Generated by roxygen2 (4.0.1): do not edit by hand
\name{landscape}
\alias{landscape}
\alias{landscape.empty}
\title{Fitness Landscape}
\usage{
landscape(trait, values, p, schedule, strategy = new(Strategy))

landscape.empty(trait, values, p, schedule, strategy = new(Strategy))
}
\arguments{
\item{trait}{Name of the single trait to change}

\item{values}{Vector of trait values to compute mutant fitness}

\item{p}{Parameters object.  Needs to contain residents with their
incoming seed rain.}

\item{schedule}{Schedule of times for the residents.}
}
\value{
Vector with the output seed rain.  Mutants have an
arbitrary seed rain of zero, so this is the rate of seed
production per capita.
}
\description{
Construct a fitness landscape.  Currently only works in one
dimension, and with all other parameters set to their defaults,
and at predefined values only rather than adaptively (i.e., not
actually useful yet).
}
\details{
Will change...
}
\author{
Rich FitzJohn
}

