\name{last}
\alias{first}
\alias{first<-}
\alias{last}
\alias{last<-}
\title{Get Last Element}
\usage{
  last(x)

  last(x, value) <- value

  first(x)

  first(x, value) <- value
}
\arguments{
  \item{x}{An object that can be subset with \code{[[}}
}
\description{
  Get last element from an object

  ##'
}
\author{
  Rich FitzJohn
}

