% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lcp_whole_plant}
\alias{lcp_whole_plant}
\title{Whole plant light compensation point}
\usage{
lcp_whole_plant(p)
}
\arguments{
\item{p}{A \code{Plant}, with strategy, height, etc set.}
}
\description{
Compute the whole plant light compensation point for a single
plant.
}
\author{
Rich FitzJohn
}

