\name{load.reference.output}
\alias{load.reference.output}
\title{Load Reference Output}
\usage{
  load.reference.output(path)
}
\arguments{
  \item{path}{Path (relative or absolute) to directory
  containing output.}
}
\value{
  A list with elements (undocumented at present).
}
\description{
  Load reference output from the previous version of the
  model.
}
\details{
  This loads the files created by the Falster version of
  the EBT model into R.  There are a lot of different
  files, one per variable.  Eventually, we will have
  functions for working with the output and getting the
  things ready for comparison.
}
\author{
  Rich FitzJohn
}
\keyword{internal}

