% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{make_ebt_integrate}
\alias{make_ebt_integrate}
\title{Integrate EBT variables}
\usage{
make_ebt_integrate(obj)
}
\arguments{
\item{ebt}{An object from \code{run_ebt} or \code{run_ebt_collect}}
}
\description{
Create a function that allows integrating aggregate properties of
the EBT system.
}
\details{
The workflow here is to run an EBT to create an EBT by running
\code{run_ebt}, or a set of data from \code{run_ebt_collect} and
then reconstitute all the intermediate bits of data so that an any
variable that \code{PlantPlus} tracks can be integrated out.
Because the pre-processing step is reasonably slow, this function
returns a function that takes a variable name and integrates it.
}

