% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/strategy_support.R
\name{environment_type}
\alias{environment_type}
\alias{make_environment}
\title{Make environment objects for a strategy}
\usage{
environment_type(type)

make_environment(type = NULL, parameters = NULL, ...)
}
\arguments{
\item{type}{Any strategy name as a string, e.g.: \code{"FF16"}.}

\item{parameters}{a object}
}
\description{
Make environment objects for a strategy
}
