% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_support.R
\name{make_equilibrium_runner}
\alias{make_equilibrium_runner}
\title{Make the equilibrium runner}
\usage{
make_equilibrium_runner(p, ctrl)
}
\arguments{
\item{p}{A Parameters object}

\item{ctrl}{Control object}
}
\value{
offspring_production
}
\description{
Make the equilibrium runner
}
