% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16w.R
\name{make_FF16w_parameters}
\alias{make_FF16w_parameters}
\title{Make a parameter tibble for use in strategy objects}
\usage{
make_FF16w_parameters(
  p0 = FF16w_Parameters(),
  lma = p0$strategy_default$lma,
  rho = p0$strategy_default$rho,
  hmat = p0$strategy_default$hmat,
  omega = p0$strategy_default$omega,
  theta = p0$strategy_default$theta,
  a_l1 = p0$strategy_default$a_l1,
  a_l2 = p0$strategy_default$a_l2,
  a_r1 = p0$strategy_default$a_r1,
  a_b1 = p0$strategy_default$a_b1,
  r_r = p0$strategy_default$r_r,
  a_y = p0$strategy_default$a_y,
  a_bio = p0$strategy_default$a_bio,
  k_b = p0$strategy_default$k_b,
  k_r = p0$strategy_default$k_r,
  a_f1 = p0$strategy_default$a_f1,
  a_f2 = p0$strategy_default$a_f2,
  S_D = p0$strategy_default$S_D,
  a_d0 = p0$strategy_default$a_d0,
  a_dG1 = p0$strategy_default$a_dG1,
  k_I = p0$strategy_default$k_I,
  vcmax_25 = p0$strategy_default$vcmax_25,
  K_s = p0$strategy_default$K_s,
  beta1 = p0$strategy_default$beta1,
  beta2 = p0$strategy_default$beta2,
  jmax_25 = p0$strategy_default$jmax_25,
  a = p0$strategy_default$a,
  curv_fact_elec_trans = p0$strategy_default$curv_fact_elec_trans,
  curv_fact_colim = p0$strategy_default$curv_fact_colim,
  nmass_s = p0$strategy_default$nmass_s,
  nmass_b = p0$strategy_default$nmass_b,
  nmass_r = p0$strategy_default$nmass_r,
  dmass_dN = p0$strategy_default$dmass_dN,
  var_sapwood_volume_cost = p0$strategy_default$var_sapwood_volume_cost
)
}
\arguments{
\item{p0}{}

\item{lma}{}

\item{rho}{}

\item{hmat}{}

\item{omega}{}

\item{theta}{}

\item{a_l1}{}

\item{a_l2}{}

\item{a_r1}{}

\item{a_b1}{}

\item{r_r}{}

\item{a_y}{}

\item{a_bio}{}

\item{k_b}{}

\item{k_r}{}

\item{a_f1}{}

\item{a_f2}{}

\item{S_D}{}

\item{a_d0}{}

\item{a_dG1}{}

\item{k_I}{}

\item{vcmax_25}{}

\item{K_s}{}

\item{beta1}{}

\item{beta2}{}

\item{jmax_25}{}

\item{a}{}

\item{curv_fact_elec_trans}{}

\item{curv_fact_colim}{}

\item{nmass_s}{}

\item{nmass_b}{}

\item{nmass_r}{}

\item{dmass_dN}{}

\item{c}{}

\item{hk_s}{}
}
\description{
Make a parameter tibble for use in strategy objects
}
