% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16w.R
\name{make_leaf}
\alias{make_leaf}
\title{Wrapper function to create a leaf object with built-in values}
\usage{
make_leaf(
  ff16w_params = make_FF16w_parameters(),
  ff16w_env = FF16w_make_environment(),
  height = 1,
  ff16w_hyper = list(lma_0 = 0.1978791),
  control = Control()
)
}
\arguments{
\item{ff16w_params}{Traits for the leaf object with a ff16w strategy matrix, uses the default ff16w strategy if NULL}

\item{ff16w_env}{Environmental values, tibble with any of the following column names: "psi_soil", "PPFD", "atm_vpd", "ca", "leaf_temp" or an ff16_environment object}

\item{height}{Height of plants in metres (m)}
}
\value{
Returns a Leaf object
}
\description{
Wrapper function to create a leaf object with built-in values
}
