% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/reference_plant.R
\name{make_reference_plant_FF16}
\alias{make_reference_plant_FF16}
\title{Reference Growth Model}
\usage{
make_reference_plant_FF16()
}
\value{
An object with a number of methods (accessed with
\code{$}).  These will hopefully be documented at some point.
}
\description{
Create an instance of reference version of growth model FF16
}
\details{
This is an old version of the growth model, implemented by Daniel
Falster.  It does \emph{not} implement any of the actual growth.
It is used in testing code to make sure that the full model
version gives sensible results.
}
\author{
Daniel S. Falster (extra code and porting by Rich
FitzJohn)
}
