% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{make_scm}
\alias{make_scm}
\title{Construct an SCM}
\usage{
make_scm(p, env, ctrl, state = NULL)
}
\arguments{
\item{p}{Parameters object}

\item{env}{Environment object (defaults to FF16_Environment)}

\item{ctrl}{Control object}

\item{state}{List object with initial cohorts for each species}
}
\description{
Functions for reconstructing a Patch from an SCM
}
