% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{make_scm_integrate}
\alias{make_scm_integrate}
\title{Integrate SCM variables}
\usage{
make_scm_integrate(obj)
}
\arguments{
\item{scm}{An object from \code{run_scm} or \code{run_scm_collect}}
}
\description{
Create a function that allows integrating aggregate properties of
the SCM system.
}
\details{
The workflow here is to run an SCM to create an SCM by running
\code{run_scm}, or a set of data from \code{run_scm_collect} and
then reconstitute all the intermediate bits of data so that an any
variable that \code{PlantPlus} tracks can be integrated out.
Because the pre-processing step is reasonably slow, this function
returns a function that takes a variable name and integrates it.
}
