% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24t.R
\name{make_TF24t_hyperpar}
\alias{make_TF24t_hyperpar}
\title{Hyperparameters for TF24t physiological model}
\usage{
make_TF24t_hyperpar(
  lma_0 = 0.1978791,
  B_kl1 = 0.4565855,
  B_kl2 = 1.71,
  rho_0 = 608,
  B_dI1 = 0.01,
  B_dI2 = 0,
  B_hks1 = 7.5,
  B_hks2 = 0,
  B_ks1 = 0.2,
  B_ks2 = 0,
  B_rs1 = 4012,
  B_rb1 = 2 * 4012,
  B_f1 = 3,
  a_lf1 = 0.535,
  B_lf1 = 0.009,
  B_lf2 = 0.004,
  B_lf3 = 8e-04,
  B_lf4 = 21000,
  B_lf5 = 40000,
  latitude = 0,
  B_Hv1 = 0.4607063,
  B_Hv2 = -0.2,
  B_c1 = 2.04,
  B_c2 = 0
)
}
\arguments{
\item{lma_0}{Central (mean) value for leaf mass per area [kg /m2]}

\item{B_kl1}{Rate of leaf turnover at lma_0 [/yr]}

\item{B_kl2}{Scaling slope for phi in leaf turnover [dimensionless]}

\item{rho_0}{Central (mean) value for wood density [kg /m3]}

\item{B_dI1}{Rate of instantaneous mortality at rho_0 [/yr]}

\item{B_dI2}{Scaling slope for wood density in intrinsic mortality [dimensionless]}

\item{B_hks1}{Intercept for the g1_TF24 ~ rho relationship at rho_0 [dimensionless]}

\item{B_hks2}{Scaling slope for rho in the g1_TF24 relationship [dimensionless]}

\item{B_ks1}{Rate of sapwood turnover at rho_0 [/yr]}

\item{B_ks2}{Scaling slope for rho in sapwood turnover [dimensionless]}

\item{B_rs1}{CO_2 respiration per unit sapwood volume [mol / yr / m3 ]}

\item{B_rb1}{CO_2 respiration per unit sapwood volume [mol / yr / m3 ]}

\item{B_f1}{Cost of seed accessories per unit seed mass [dimensionless]}

\item{a_lf1}{intercept for empirical relationship between narea and vcmax, lma (Dong et al. 2022)}

\item{B_lf1}{Beta coefficient for empirical relationship between narea_ls ~ lma [g/m2] (Dong et al. 2022)}

\item{B_lf2}{Beta coefficient for empirical relationship between narea_lp ~ vcmax [g/m2] (Dong et al. 2022)}

\item{B_lf3}{Beta coefficient for empirical relationship between narea_lp ~ jmax [umol / m2 / s] (Dong et al. 2022)}

\item{B_lf4}{CO_2 respiration per unit structural leaf nitrogen [mol / yr / kg]}

\item{B_lf5}{CO_2 respiration per unit photosynthetic leaf nitrogen [mol / yr / kg]}

\item{latitude}{degrees from equator (0-90), used in solar model [deg]}

\item{B_Hv1}{p50 at K_s = 1 [-MPa]}

\item{B_Hv2}{Scaling slope for K_s in p50 [dimensionless]}

\item{B_c1}{Shape parameter c of the vulnerability curve at p_50 = 0 [dimensionless]}

\item{B_c2}{Scaling slope for p_50 in the vulnerability-curve shape parameter c [dimensionless]}
}
\description{
Hyperparameters for TF24t physiological model
}
