% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_transparent}
\alias{make_transparent}
\title{Make colours transparent}
\usage{
make_transparent(col, opacity = 0.5)
}
\arguments{
\item{col}{Vector of colours}

\item{opacity}{Vector of opacities}
}
\description{
Make colours transparent
}
\examples{
make_transparent("red", seq(0, 1, length.out=6))
make_transparent(c("red", "blue"), .5)
}
