% Generated by roxygen2 (4.0.2): do not edit by hand
\name{max_growth_rate}
\alias{max_growth_rate}
\title{Compute Max Growth Rate}
\usage{
max_growth_rate(trait, values, p, schedule = NULL)
}
\arguments{
\item{trait}{Name of the trait (e.g., \code{"lma"})}

\item{values}{Values to compute maximum growth rate for}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{schedule}{\code{CohortSchedule} to use, or \code{NULL} to
generate a hopefully reasonable schedule.}
}
\description{
Compute max growth rate of a given set of values for a trait.
This is the log of per-capita seed production (i.e., fitness).
}
\details{
Only works in one dimension
}
\author{
Rich FitzJohn
}

