% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/util_solve.R
\name{nlsolve}
\alias{nlsolve}
\title{Thin wrapper around nleqslv and dfsane}
\usage{
nlsolve(x, fn, tol = 1e-06, maxit = 100, solver = "nleqslv")
}
\arguments{
\item{x}{Starting point}

\item{fn}{Function to solve}

\item{tol}{Tolerance (for \code{nleqslv} this will be both
absolute and relative)}

\item{maxit}{Maximum number of iterations.  The number of function
evaluations will likely exceed this.}

\item{solver}{The solver to use.  Either "nleqslv" or "dfsane" for now.}
}
\description{
Thin wrapper around \code{nleqslv} and \code{dfsane}
}
