% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Parameters}
\alias{Parameters}
\title{Strategy parameters}
\usage{
Parameters(..., values = list(...))
}
\arguments{
\item{...,values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Strategy parameters that tune various aspects of the biological model.
}

