% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{Plant}
\alias{Plant}
\title{Plant object}
\usage{
Plant(T, E)
}
\arguments{
\item{T}{String containing class of plant strategy to create;}

\item{E}{String containing class of environmnet to create;}
}
\description{
Plant object
}
