% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/plot_patch.R
\name{plot_patch_output_across_time}
\alias{plot_patch_output_across_time}
\title{Title}
\usage{
plot_patch_output_across_time(
  patch_output_data,
  steps = c(100, 200, 250, 300, 308)
)
}
\arguments{
\item{patch_output_data}{}

\item{steps}{}
}
\value{

}
\description{
Title
}
