% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/plot_patch.R
\name{plot_size_distribution_patch}
\alias{plot_size_distribution_patch}
\title{Title}
\usage{
plot_size_distribution_patch(data_species)
}
\arguments{
\item{data_species}{}
}
\value{

}
\description{
Title
}
