% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rbind_list}
\alias{rbind_list}
\alias{cbind_list}
\title{Create matrices from lists}
\usage{
rbind_list(x)

cbind_list(x)
}
\arguments{
\item{x}{A list, or something coercable to a list}
}
\description{
Create a matrix from a list by rbinding all columns together
}
