\name{run.ebt.collect}
\alias{run.ebt.collect}
\title{Run the EBT, Collecting Output}
\usage{
  run.ebt.collect(p, sched)
}
\arguments{
  \item{p}{A Parameters object}

  \item{sched}{A CohortSchedule Object}
}
\description{
  Run the EBT model, given a Parameters and CohortSchedule
}
\details{
  This is mostly a simple wrapper around some of the EBT
  functions. Not sure if this is how we will generally want
  to do this. Consider this function liable to change.
}
\author{
  Rich FitzJohn
}

