\name{run.ebt}
\alias{run.ebt}
\title{Run EBT}
\usage{
run.ebt(p, sched)
}
\arguments{
  \item{p}{Parameters object}

  \item{sched}{CohortSchedule object}
}
\value{
A \code{EBT} object.
}
\description{
Run the EBT, returning the EBT object for interrogation
}
\details{
This is the simplest way of using the EBT, probably.
}
\author{
Rich FitzJohn
}

