% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{run_stochastic_collect}
\alias{run_stochastic_collect}
\title{Run a stochastic patch, Collecting Output}
\usage{
run_stochastic_collect(
  p,
  env = Environment(parameters = p),
  ctrl = scm_base_control(),
  random_schedule = TRUE
)
}
\arguments{
\item{p}{A \code{\link{FF16_Parameters}} object}

\item{env}{Environment object}

\item{ctrl}{Control object}

\item{random_schedule}{setting to TRUE causes algorithm to generate
a random schedule based on offspring arrival and area.}
}
\description{
Run a stochastic simulation of a patch, given a Parameters
}
\details{
This one might need to be made differently so that different
schedules can be added easily.
Not sure if this is how we will generally want to do this.
Consider this function liable to change.
}
\author{
Rich FitzJohn
}
