% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/scm_support.R
\name{scm_base_parameters}
\alias{scm_base_parameters}
\title{This used to be bundled into base control but refactoring to separate
concerns. May disappear in the future.}
\usage{
scm_base_parameters(type = "FF16", env = environment_type(type))
}
\description{
This used to be bundled into base control but refactoring to separate
concerns. May disappear in the future.
}
