% Generated by roxygen2 (4.0.0): do not edit by hand
\name{selection_gradient}
\alias{selection_gradient}
\title{Returns selection gradient in single species communities
at a range of trait values.}
\usage{
selection_gradient(trait, values, p, dx = 1e-04, log_scale = TRUE,
  seed_rain = NULL, verbose = TRUE)
}
\arguments{
\item{trait}{name of trait.}

\item{values}{vector of trait values.}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{dx=1e-4}{Interval over which derivative is calculated}

\item{log_scale=TRUE}{Determines whether derivative is taken
with respect to raw or log-transformed x values. The latter
is useful when x is log-normally distributed.}

\item{seed_rain=NULL}{(optional) Starting values for seed rain
when solving for demographic euqilibrium}

\item{verbose=TRUE}{(optional) Print values to screen}
}
\value{
a vector of values with same length as \code{values}.
}
\description{
Returns selection gradient in single species communities
at a range of trait values. This is derivative of fitness
with respect to trait value. The algorithm first solves
for the demographic attractor, using \code{seed_rain} as
a starting value.
}
\author{
Daniel Falster
}
\seealso{
selection_gradient1
}

