% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_trait_space.R
\name{solve_max_worker}
\alias{solve_max_worker}
\title{Execute finding point of maximum fitness within some range}
\usage{
solve_max_worker(bounds, f, tol = 0.001, outcome)
}
\arguments{
\item{bounds}{Two element vector specifying range within which to}

\item{f}{function passed from solve_max_fitness}

\item{tol}{Tolerance used in the optimisation}

\item{outcome}{Desired outcome, e.g. "fitness"}
}
\value{
ret
}
\description{
Execute finding point of maximum fitness within some range
}
