% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_00.R
\name{ST_00_Plant}
\alias{ST_00_Plant}
\alias{lcp_whole_plant.Plant<ST_00>}
\alias{ST_00_Cohort}
\alias{ST_00_Species}
\alias{ST_00_Parameters}
\alias{ST_00_Patch}
\alias{ST_00_SCM}
\alias{ST_00_StochasticSpecies}
\alias{ST_00_StochasticPatch}
\alias{ST_00_StochasticPatchRunner}
\title{Create a ST_00 Plant or Cohort}
\usage{
ST_00_Plant(s = ST_00_Strategy())

\method{lcp_whole_plant}{`Plant<ST_00>`}(p, ...)

ST_00_Cohort(s = ST_00_Strategy())

ST_00_Species(s = ST_00_Strategy())

ST_00_Parameters()

ST_00_Patch(p)

ST_00_SCM(p)

ST_00_StochasticSpecies(s = ST_00_Strategy())

ST_00_StochasticPatch(p)

ST_00_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{ST_00_Strategy}} object}

\item{p}{A \code{Parameters<ST_00,ST_00_Env>} object}

\item{...}{Arguments!}
}
\description{
Create a ST_00 Plant or Cohort
}
\examples{
pl <- ST_00_Plant()
pl$height
}
