% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/st_00.R
\name{ST_00_Environment}
\alias{ST_00_Environment}
\alias{ST_00_make_environment}
\title{ST_00_Environment object}
\usage{
ST_00_Environment(disturbance_mean_interval, seed_rain, k_I, control)

ST_00_make_environment(p)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{k_I}{light extinction coefficient}

\item{control}{A Control object}

\item{p}{A Parameters object}
}
\description{
ST_00_Environment object
}
