% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{ST_00_Strategy}
\alias{ST_00_Strategy}
\title{Strategy parameters}
\usage{
ST_00_Strategy(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Strategy parameters that tune various aspects of the biological model.
}
