% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_model.R
\name{sun_rise}
\alias{sun_rise}
\title{Title Calculate the hour of sun rise}
\usage{
sun_rise(decimal_day_time, latitude)
}
\arguments{
\item{decimal_day_time}{Day of the year in decimal time}

\item{latitude}{Latitude in degrees}
}
\value{
Hour of sunrise (hr)
}
\description{
Title Calculate the hour of sun rise
}
\examples{
sun_rise(90, -27)

}
