% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_Individual}
\alias{TF24_Individual}
\alias{TF24_Node}
\alias{TF24_Species}
\alias{TF24_Parameters}
\alias{TF24_Patch}
\alias{TF24_SCM}
\alias{TF24_StochasticSpecies}
\alias{TF24_StochasticPatch}
\alias{TF24_StochasticPatchRunner}
\title{Create a TF24 Plant or Node}
\usage{
TF24_Individual(s = TF24_Strategy())

TF24_Node(s = TF24_Strategy())

TF24_Species(s = TF24_Strategy())

TF24_Parameters()

TF24_Patch(p)

TF24_SCM(p)

TF24_StochasticSpecies(s = TF24_Strategy())

TF24_StochasticPatch(p)

TF24_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{TF24_Strategy}} object}

\item{p}{A \code{Parameters<TF24,TF24_Env>} object}
}
\description{
Create a TF24 Plant or Node
}
\examples{
pl <- TF24_Individual()
pl$height
}
