% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{TF24_Environment}
\alias{TF24_Environment}
\title{TF24 Environment}
\usage{
TF24_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths,
  delta_z,
  soil_moist_sat,
  K_sat,
  a_psi,
  n_psi,
  b_infil
)
}
\arguments{
\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}

\item{soil_number_of_depths}{Number of soil layers to include}

\item{delta_z}{Thickness of soil layers}

\item{soil_moist_sat}{Soil moisture at saturation}

\item{K_sat}{Saturated hydraulic conductivity}

\item{a_psi}{Coefficient for soil water potential}

\item{n_psi}{Exponent for soil water potential}

\item{b_infil}{Coefficient for infiltration}
}
\description{
This environment is used for the TF24 strategy, which includes soil moisture and hydraulic
conductivity parameters.
}
\details{
Create an TF24_Environment object.
}
