% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_fixed_environment}
\alias{TF24_fixed_environment}
\title{Construct a fixed environment for a TF24 model}
\usage{
TF24_fixed_environment(e = 1, height_max = 150, ...)
}
\arguments{
\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{maximum possible height in environment}

\item{...}{Additional parameters to be passed to \code{FF16_make_environment}}
}
\description{
Construct a fixed environment for a TF24 model
}
