% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tf24.R
\name{TF24_hyperpar}
\alias{TF24_hyperpar}
\title{Hyperparameter function for TF24 physiological model}
\usage{
TF24_hyperpar(m, s, filter = TRUE)
}
\arguments{
\item{m}{A matrix of trait values, as returned by \code{trait_matrix}}

\item{s}{A strategy object}

\item{filter}{A flag indicating whether to filter columns. If TRUE, any numbers
that are within eps of the default strategy are not replaced.}
}
\description{
Hyperparameter function for TF24 physiological model
}
