% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_make_environment}
\alias{TF24_make_environment}
\title{: Helper functions to create an TF24_Environment object. Useful for running individuals}
\usage{
TF24_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE,
  soil_number_of_depths = 1,
  soil_initial_state = rep(0.25, soil_number_of_depths),
  rainfall = 1,
  atm_vpd = 1,
  ca = 40,
  leaf_temp = 25,
  atm_o2_kpa = 21,
  atm_kpa = 100.5,
  delta_z = 0.1,
  soil_moist_sat = 0.453,
  K_sat = 440.628,
  a_psi = 8.7,
  n_psi = 4.8,
  b_infil = 8
)
}
\arguments{
\item{light_availability_spline_tol}{Error tolerance of adpative spline method. Deafult is 1e-4.}

\item{light_availability_spline_nbase}{Parameter used in adaptive spline method. Default is 17.}

\item{light_availability_spline_max_depth}{Parameter used in adaptive spline method. Default is 16.}

\item{soil_initial_state}{Initial state of the soil moisture, a vector of length `soil_number_of_depths`}
}
\description{
: Helper functions to create an TF24_Environment object. Useful for running individuals
}
