% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{TF24_Pars}
\alias{TF24_Pars}
\title{TF24 strategy parameters}
\usage{
TF24_Pars(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Biological trait/parameters for the TF24 strategy. Held as the nested
\code{pars} element of a \code{\link{TF24_Strategy}} (access as
\code{s$pars$lma}).
}
