% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{TF24_Strategy}
\alias{TF24_Strategy}
\title{Strategy parameters}
\usage{
TF24_Strategy(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Strategy parameters that tune various aspects of the biological model.
}
