% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_test_environment}
\alias{TF24_test_environment}
\title{Create a test environment for TF24 startegy. Only used in testing}
\usage{
TF24_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)
}
\arguments{
\item{height}{top height of environment object}

\item{n}{number of points}

\item{light_env}{function for light environment in test object}

\item{n_strategies}{number of strategies for test environment}
}
\description{
Create a test environment for TF24 startegy. Only used in testing
}
