% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24f.R
\name{TF24f_Individual}
\alias{TF24f_Individual}
\title{Create a TF24f Plant or Node}
\usage{
TF24f_Individual(s = TF24f_Strategy())
}
\arguments{
\item{s}{A \code{\link{TF24f_Strategy}} object}
}
\description{
Create a TF24f Plant or Node
}
\examples{
pl <- TF24f_Individual()
pl$height
}
