% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24f.R
\name{TF24f_Parameters}
\alias{TF24f_Parameters}
\title{Setup an a TF24f system with  default or specified parameters}
\usage{
TF24f_Parameters(...)
}
\arguments{
\item{...}{Arguments to be passed to the model constructor. These include

  *`patch_area`: Area of idnividfual patch. Only relevant for stochastic model. Default is 1.0m2.
  *`max_patch_lifetime`: The maximum time in years we want to simulate
  *`strategies`: A list of stratgies to simulate. The default is an empty list.
  *`strategy_default`: Values for the default startegy. The default values are those specified in the C++ code for the model.
  *`node_schedule_times_default`: Default vector of times at which to introduce nodes. The default is chosen to have close spacing at the start of the simulation.
  *`node_schedule_times`: A list with each element containing the vector of times we want to introduce nodes for each strategy. The default is an empty list.
  *`ode_times`: A vector of patch ages we want the ode solver to stop at}
}
\description{
Setup an a model system with default or specified parameters.
}
\examples{
p1 <- TF24f_Parameters()
p2 <- TF24f_Parameters(max_patch_lifetime = 10.0)
}
