% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24t.R
\name{TF24t_Individual}
\alias{TF24t_Individual}
\title{Create a TF24t Plant or Node}
\usage{
TF24t_Individual(s = TF24t_Strategy())
}
\arguments{
\item{s}{A \code{\link{TF24t_Strategy}} object}
}
\description{
Create a TF24t Plant or Node
}
\examples{
pl <- TF24t_Individual()
pl$height
}
