% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_outputs.R
\name{tidy_env}
\alias{tidy_env}
\title{Turn `env` component of solver output into a tidy data object}
\usage{
tidy_env(env)
}
\arguments{
\item{env}{a list, the `env` component of solver output.}
}
\value{
a tibble describing the environment in a patch
}
\description{
Turn `env` component of solver output into a tidy data object
}
