% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_outputs.R
\name{tidy_individual}
\alias{tidy_individual}
\title{Turn `results` of plant solver, when solving individuals into a tidy data object}
\usage{
tidy_individual(results)
}
\arguments{
\item{results}{plant solver output.}
}
\value{
a tibble whose columns provide metrics on each individual over time
}
\description{
Turn `results` of plant solver, when solving individuals into a tidy data object
}
