% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/water.R
\name{Water_Individual}
\alias{Water_Individual}
\alias{Water_Cohort}
\alias{Water_Species}
\alias{Water_Parameters}
\alias{Water_Patch}
\alias{Water_SCM}
\alias{Water_StochasticSpecies}
\alias{Water_StochasticPatch}
\alias{Water_StochasticPatchRunner}
\title{Create a Water Plant or Cohort}
\usage{
Water_Individual(s = Water_Strategy())

Water_Cohort(s = Water_Strategy())

Water_Species(s = Water_Strategy())

Water_Parameters()

Water_Patch(p)

Water_SCM(p)

Water_StochasticSpecies(s = Water_Strategy())

Water_StochasticPatch(p)

Water_StochasticPatchRunner(p)
}
\arguments{
\item{s}{A \code{\link{Water_Strategy}} object}

\item{p}{A \code{Parameters<Water,FF16_Env>} object}
}
\description{
Create a Water Plant or Cohort
}
\examples{
pl <- Water_Individual()
pl$height
}
