% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/water.R
\name{Water_test_environment}
\alias{Water_test_environment}
\title{Create a test environment for Water startegy}
\usage{
Water_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)
}
\arguments{
\item{height}{top height of environment object}

\item{n}{number of points}

\item{light_env}{function for light environment in test object}

\item{n_strategies}{number of strategies for test environment}

\item{birth_rate}{birth_rate for test environment}
}
\description{
This makes a pretend light environment over the plant height,
slightly concave up, whatever.
}
\examples{
environment <- Water_test_environment(10)
}
