% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Weibull_Disturbance_Regime}
\alias{Weibull_Disturbance_Regime}
\title{Weibull disturbance regime
Weibull_Disturbance_Regime control object}
\usage{
Weibull_Disturbance_Regime(max_patch_lifetime)
}
\description{
The frequency of disturbance in a meta-population of patches follows a Weibull distribution
}
