% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/scm_support.R
\name{assembly_parameters}
\alias{assembly_parameters}
\title{Helper function for creating parameter objects}
\usage{
assembly_parameters(
  ...,
  pars = NULL,
  base_parameters_fn = scm_base_parameters,
  make_hyperpar_fn = make_FF16_hyperpar
)
}
\arguments{
\item{...}{Named set of parameters}

\item{pars}{A list of parameters}

\item{base_parameters_fn}{Function for creating base parameter set (default scm_base_parameters)}

\item{make_hyperpar_fn}{Function for creating hyperparameterisation (default make_FF16_hyperpar)}
}
\description{
Helper function for creating parameter objects suitable for an
assembly.
}
