% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bounds}
\alias{bounds}
\alias{bounds_infinite}
\alias{check_bounds}
\title{Trait bounds}
\usage{
bounds(...)

check_bounds(bounds, finite = FALSE)

bounds_infinite(trait_names)
}
\arguments{
\item{...}{Named list, each element of which is a 2-element
numeric vector of lower and upper bounds.}

\item{bounds}{A set of bounds}

\item{finite}{Logical indicating if bounds must be finite}

\item{trait_names}{Character vector of trait names}
}
\description{
Helper function for making bounds
}
\examples{
bounds(lma=c(0.01, 10))
bounds(lma=c(0.01, 10), rho=c(1, 1000))
}

