% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/fitness_support.R
\name{bounds}
\alias{bounds}
\alias{check_bounds}
\alias{bounds_infinite}
\alias{check_point}
\title{Trait bounds}
\usage{
bounds(...)

check_bounds(bounds, finite = FALSE)

bounds_infinite(trait_names)

check_point(x, bounds)
}
\arguments{
\item{...}{Named list, each element of which is a 2-element
numeric vector of lower and upper bounds.}

\item{bounds}{A set of bounds}

\item{finite}{Logical indicating if bounds must be finite}

\item{trait_names}{Character vector of trait names}

\item{x}{a point to detect if it lies within bounds}
}
\description{
Helper function for making bounds
}
\examples{
bounds(lma=c(0.01, 10))
bounds(lma=c(0.01, 10), rho=c(1, 1000))
}
