% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_schedule}
\alias{build_schedule}
\title{Build Cohort Schedule}
\usage{
build_schedule(p, schedule = NULL)
}
\arguments{
\item{p}{Parameters object}

\item{schedule}{An optional CohortSchedule, with a schedule to
start from.  If \code{NULL} (the default) a reasonable default
will be made using \code{\link{default_cohort_schedule}}}
}
\value{
A CohortSchedule object
}
\description{
Build an appropriately refined schedule.
}
\details{
There are control options (within the \code{Parameters} object)
that affect how this function runs, in particular
\code{schedule_nsteps} and \code{schedule_eps} control how refined
the schedule will end up, and \code{schedule_verbose} controls if
details are printed to the screen during construction.
}
\author{
Rich FitzJohn
}

