% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_schedule.R
\name{build_schedule}
\alias{build_schedule}
\title{Build Cohort Schedule}
\usage{
build_schedule(
  p,
  env = make_environment(parameters = p),
  ctrl = scm_base_control(),
  splines = NULL,
  n_init = 10
)
}
\arguments{
\item{p}{Parameters object}

\item{ctrl}{Control object}

\item{splines}{A list of interpolating spine functions describing the
initial size distribution for the patch (generated by \code{\link{init_spline}})}

\item{n_init}{the number of initial cohorts to start with (default = 10)}
}
\value{
A Parameters object, with schedule components set.  The
output offspring produced is also available as an attribute
\code{birth_rate}.
}
\description{
Build an appropriately refined schedule.
}
\details{
There are control options (within the \code{Parameters} object)
that affect how this function runs, in particular
\code{schedule_nsteps} and \code{schedule_eps} control how refined
the schedule will end up, and \code{schedule_verbose} controls if
details are printed to the screen during construction.
}
\author{
Rich FitzJohn
}
