% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_inviable}
\alias{check_inviable}
\title{Check low-abundance strategies for viability}
\usage{
check_inviable(p)
}
\arguments{
\item{p}{A Parameters object}

\item{schedule_default}{Schedule to use by default}

\item{schedule_initial}{Schedule to start from}
}
\description{
Check low-abundance strategies for viability.
}

