% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cobweb}
\alias{cobweb}
\title{Cobweb Plot}
\usage{
cobweb(m, ...)
}
\arguments{
\item{m}{Two column matrix}

\item{...}{Additional parameters passed to \code{lines}}
}
\description{
Add points and lines for a cobweb plot showing convergence to an
equilibrium.
}
\details{
Assumed (unchecked) that there are two columns corresponding to
"from" and "to.
}
\author{
Rich FitzJohn
}

