% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Control}
\alias{Control}
\title{Control parameters}
\usage{
Control(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Control parameters that tune various aspects of the numerical solvers.
}
