% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default_cohort_schedule}
\alias{default_cohort_schedule}
\title{Generate Default Cohort Schedule}
\usage{
default_cohort_schedule(p, t_max = NULL)
}
\arguments{
\item{p}{A Parameters object}

\item{t_max}{Maximum time to run schedule for.  If omitted (the
usual case) then time proceeds until there is a "very small"
chance that the patch is still extant, based on the control
parameter \code{schedule_default_patch_survival}.}
}
\description{
Generate default cohort schedule.  This has times compressed
towards young patch ages, which usually need finer resolution than
later patche ages.  Currently there is no support for arbitrary
times: just build the schedule manually for now.
}
\author{
Rich FitzJohn
}

