% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{ebt_base_parameters}
\alias{ebt_base_parameters}
\title{Sensible, fast (ish) EBT parameters}
\usage{
ebt_base_parameters()
}
\description{
Hopefully sensible set of parameters for use with the EBT.  Turns
accuracy down a bunch, makes it noisy, sets up the
hyperparameterisation that we most often use.
}
\author{
Rich FitzJohn
}

