% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/ff16.R
\name{Environment}
\alias{Environment}
\alias{make_environment}
\title{Environment object}
\usage{
Environment(disturbance_mean_interval, seed_rain, control)

make_environment(p)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{control}{A Control object}

\item{p}{A Parameters object}
}
\description{
Environment object
}
