% Generated by roxygen2 (4.0.2): do not edit by hand
\name{equilibrium_seed_rain}
\alias{equilibrium_seed_rain}
\title{Run system to seed rain equilibrium}
\usage{
equilibrium_seed_rain(p)
}
\arguments{
\item{p}{A \code{\link{Parameters}} object}
}
\value{
A Parameters object, with seed rain and cohort schedule
elements set.
}
\description{
Run system to seed rain equilibrium
}
\author{
Rich FitzJohn
}

