% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/equilibrium.R
\name{equilibrium_seed_rain}
\alias{equilibrium_seed_rain}
\title{Run system to seed rain equilibrium}
\usage{
equilibrium_seed_rain(p)
}
\arguments{
\item{p}{A \code{Parameters} object}
}
\value{
A Parameters object, with seed rain and cohort schedule
elements set.
}
\description{
Run system to seed rain equilibrium
}
\author{
Rich FitzJohn
}
