% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand Parameters to include mutants}
\usage{
expand_parameters(trait_matrix, p, mutant = TRUE)
}
\arguments{
\item{trait_matrix}{A matrix of traits corresponding to the
new types to introduce.}

\item{p}{A \code{Parameters} object.}
}
\description{
Expand Parameters to include mutants.  All mutants get the same
schedule, equal to all the unique times that any resident was
introduced (if mutants) or the default schedule (if residents).
This results in more work than is really needed, but should be
reasonable most of the time.
}
\author{
Rich FitzJohn
}

