% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand Parameters to Include Mutants}
\usage{
expand_parameters(trait, values, p, mutant = TRUE)
}
\arguments{
\item{trait}{Name of the single trait to change}

\item{values}{Vector of trait values for the mutants}

\item{p}{Parameters object}

\item{mutant}{Are these mutant strategies? (Default: yes they
are).}
}
\description{
Expand parameters to include mutants.  All mutants are built off
the default strategy.
}
\author{
Rich FitzJohn
}

