% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_model.R
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand Parameters to include mutants}
\usage{
expand_parameters(
  trait_matrix,
  p,
  hyperpar = param_hyperpar(p),
  birth_rate_list = 1
)
}
\arguments{
\item{trait_matrix}{A matrix of traits corresponding to the
new types to introduce.}

\item{p}{A \code{Parameters} object.}

\item{hyperpar}{Hyperparameter function to use. By default links
to standard function for this strategy type.}

\item{birth_rate_list}{List object with birth rates for each species in
x. Birth rates can take the form of a scalar (constant) or a vector.
In either case birth rates are set as \code{strategy$birth_rate_y}, however
varying birth rates will also have \code{strategy$birth_rate_x} and}
}
\description{
Expand Parameters to include mutants.  All mutants get the same
schedule, equal to all the unique times that any resident was
introduced (if mutants) or the default schedule (if residents).
This results in more work than is really needed, but should be
reasonable most of the time.
}
\author{
Rich FitzJohn
}
