% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_model.R
\name{expand_parameters}
\alias{expand_parameters}
\alias{mutant_parameters}
\title{Setup parameters to run resindets or mutants}
\usage{
expand_parameters(
  trait_matrix,
  p,
  hyperpar = param_hyperpar(p),
  birth_rate_list = 1,
  keep_existing_strategies = TRUE
)

mutant_parameters(..., keep_existing_strategies = FALSE)
}
\arguments{
\item{trait_matrix}{A matrix of traits corresponding to the
new types to introduce.}

\item{p}{A \code{Parameters} object.}

\item{hyperpar}{Hyperparameter function to use. By default links
to standard function for this strategy type.}

\item{birth_rate_list}{List object with birth rates for each species in}

\item{keep_existing_strategies}{Should existing residents be retained
x. Birth rates can take the form of a scalar (constant) or a vector.
In either case birth rates are set as \code{strategy$birth_rate_y}, however
varying birth rates will also have \code{strategy$birth_rate_x} and}
}
\description{
The functions expand_parameters and mutant_parameters convert trait values into parametr objects for the model. By default, expand_parameters adds an extra strategy to existing.
}
\author{
Rich FitzJohn
}
