% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_model.R
\name{expand_parameters}
\alias{expand_parameters}
\title{Expand Parameters to include mutants}
\usage{
expand_parameters(trait_matrix, p, mutant = TRUE)
}
\arguments{
\item{trait_matrix}{A matrix of traits corresponding to the
new types to introduce.}

\item{p}{A \code{Parameters} object.}

\item{mutant}{Create new types as \emph{mutants}?  These will have
no effect on other plants in the community (i.e. have zero
density).}
}
\description{
Expand Parameters to include mutants.  All mutants get the same
schedule, equal to all the unique times that any resident was
introduced (if mutants) or the default schedule (if residents).
This results in more work than is really needed, but should be
reasonable most of the time.
}
\author{
Rich FitzJohn
}
