% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R, R/k93.R, R/strategy_support.R,
%   R/tf24.R, R/tf24f.R, R/tf24t.R
\name{FF16_expand_state}
\alias{FF16_expand_state}
\alias{K93_expand_state}
\alias{expand_state}
\alias{TF24_expand_state}
\alias{TF24f_expand_state}
\alias{TF24t_expand_state}
\title{Add additional state variables to the species component in output of a model.}
\usage{
FF16_expand_state(results)

K93_expand_state(results)

expand_state(results)

TF24_expand_state(results)

TF24f_expand_state(results)

TF24t_expand_state(results)
}
\arguments{
\item{results}{from `tidy_patch`}
}
\value{
similar format to input, but with additional columns for additional state variables
}
\description{
Add additional state variables to the species component in output of a model.
}
