% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fast_control}
\alias{fast_control}
\title{Fast Control Defaults}
\usage{
fast_control()
}
\value{
A `list` of values to be passed into
\code{Control$set_parameters}
}
\description{
Parameters for running the simulations more quickly (but less
accurately) than the defaults.  Used in a number of places.
}
\author{
Rich FitzJohn
}

