% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{fast_control}
\alias{fast_control}
\title{Fast Control Defaults}
\usage{
fast_control(base = Control())
}
\arguments{
\item{base}{An optional \code{Control} object.  If omitted, the
defaults are used.}
}
\value{
A Control object with parameters set.
}
\author{
Rich FitzJohn
}
