% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R
\name{FF16_Plant}
\alias{FF16_Plant}
\alias{FF16_Cohort}
\alias{FF16_Species}
\alias{FF16_Parameters}
\alias{FF16_Patch}
\alias{FF16_SCM}
\alias{FF16_StochasticSpecies}
\alias{FF16_StochasticPatch}
\alias{FF16_StochasticPatchRunner}
\alias{FF16_PlantPlus}
\title{Create a FF16 Plant or Cohort}
\usage{
FF16_Plant(s = FF16_Strategy())

FF16_Cohort(s = FF16_Strategy())

FF16_Species(s = FF16_Strategy())

FF16_Parameters(...)

FF16_Patch(p)

FF16_SCM(p)

FF16_StochasticSpecies(s = FF16_Strategy())

FF16_StochasticPatch(p)

FF16_StochasticPatchRunner(p)

FF16_PlantPlus(s = FF16_Strategy())
}
\arguments{
\item{s}{A \code{\link{FF16_Strategy}} object}

\item{...}{Arguments!}

\item{p}{A \code{Parameters<FF16>} object}
}
\description{
Create a FF16 Plant or Cohort
}
\examples{
pl <- FF16_Plant()
pl$height
}
