% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{K93_Environment}
\title{Create an FF16_Environment object.}
\usage{
FF16_Environment(light_availability_spline_rescale_usually)

K93_Environment()
}
\arguments{
\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}
}
\description{
Create an FF16_Environment object.
}
