% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R, r/ff16.R, r/ff16r.R, r/water.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{FF16_make_environment}
\alias{FF16_fixed_environment}
\alias{FF16r_make_environment}
\alias{Water_make_environment}
\alias{Water_fixed_environment}
\title{FF16_Environment object}
\usage{
FF16_Environment(control)

FF16_make_environment(ctrl = scm_base_control())

FF16_fixed_environment(e = 1, ctrl = scm_base_control(), height_max = 150)

FF16r_make_environment(ctrl = scm_base_control())

Water_make_environment(ctrl = scm_base_control())

Water_fixed_environment(e = 1, ctrl = scm_base_control(), height_max = 150)
}
\arguments{
\item{control}{A Control object}

\item{ctrl}{Control object}

\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}
}
\description{
FF16_Environment object

Construct a fixed environment for FF16 strategy

Construct a fixed environment for Water strategy
}
