% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/ff16.R, R/ff16r.R, R/ff16w.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{FF16_make_environment}
\alias{FF16_fixed_environment}
\alias{FF16r_make_environment}
\alias{FF16w_make_environment}
\alias{FF16w_fixed_environment}
\title{FF16_Environment object}
\usage{
FF16_Environment(canopy_rescale_usually, soil_number_of_depths)

FF16_make_environment(
  canopy_light_tol = 1e-04,
  canopy_light_nbase = 17,
  canopy_light_max_depth = 16,
  canopy_rescale_usually = TRUE
)

FF16_fixed_environment(e = 1, height_max = 150)

FF16r_make_environment(canopy_rescale_usually = TRUE)

FF16w_make_environment(
  canopy_light_tol = 1e-04,
  canopy_light_nbase = 17,
  canopy_light_max_depth = 16,
  canopy_rescale_usually = TRUE,
  soil_number_of_depths = 1,
  soil_initial_state = 0.5,
  rainfall = 1,
  vpd = 1,
  co2 = 40,
  leaf_temp = 25,
  atm_o2 = 21,
  atm_kpa = 100.5
)

FF16w_fixed_environment(e = 1, height_max = 150)
}
\arguments{
\item{canopy_rescale_usually}{}

\item{soil_number_of_depths}{the number of soil layers}

\item{canopy_light_tol}{}

\item{canopy_light_nbase}{}

\item{canopy_light_max_depth}{}

\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}

\item{soil_initial_state}{initial soil moisture (m3 m^-3)}

\item{rainfall}{Rainfall (m). Can be a constant value or a list (x = time, y = rainfall)}

\item{vpd}{Vapour pressure deficit (kPa). Can be a constant value or a list (x = time, y = VPD)}

\item{co2}{CO2 partial pressure (Pa). Can be a constant value or a list (x = time, y = co2)}
}
\description{
FF16_Environment object

Construct a fixed environment for FF16 strategy

Construct a fixed environment for FF16w strategy
}
