% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/ff16.R, R/ff16r.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{FF16_fixed_environment}
\alias{FF16r_make_environment}
\title{FF16_Environment object}
\usage{
FF16_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths
)

FF16_fixed_environment(e = 1, height_max = 150)

FF16r_make_environment(light_availability_spline_rescale_usually = TRUE)
}
\arguments{
\item{light_availability_spline_rescale_usually}{turns on environment rescaling (default = TRUE)}

\item{soil_number_of_depths}{Number of soil layers to include}

\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}
}
\description{
FF16_Environment object

Construct a fixed environment for FF16 strategy
}
