% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/ff16.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{make_environment}
\title{FF16_Environment object}
\usage{
FF16_Environment(disturbance_mean_interval, seed_rain, k_I, control)

make_environment(p)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{k_I}{light extinction coefficient}

\item{control}{A Control object}

\item{p}{A Parameters object}
}
\description{
FF16_Environment object
}
