% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R, R/k93.R
\name{FF16_fixed_environment}
\alias{FF16_fixed_environment}
\alias{K93_fixed_environment}
\title{Construct a fixed environment for a model}
\usage{
FF16_fixed_environment(e = 1, height_max = 150, ...)

K93_fixed_environment(e = 1, height_max = 300, ...)
}
\arguments{
\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{maximum possible height in environment}

\item{...}{Additional parameters to be passed to \code{FF16_make_environment}}
}
\description{
Construct a fixed environment for a model
}
