% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R, r/ff16w.R, r/k93.R, r/tf24.R
\name{FF16_make_environment}
\alias{FF16_make_environment}
\alias{FF16w_make_environment}
\alias{K93_make_environment}
\alias{TF24_make_environment}
\title{create a model Environment object}
\usage{
FF16_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)

FF16w_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE,
  soil_number_of_depths = 1,
  soil_initial_state = 0,
  rainfall = 1
)

K93_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)

TF24_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE,
  soil_number_of_depths = 1,
  soil_initial_state = rep(0.25, soil_number_of_depths),
  rainfall = 1,
  atm_vpd = 1,
  ca = 40,
  leaf_temp = 25,
  atm_o2_kpa = 21,
  atm_kpa = 100.5,
  delta_z = 0.1,
  soil_moist_sat = 0.453,
  K_sat = 440.628,
  a_psi = 8.7,
  n_psi = 4.8,
  b_infil = 8
)
}
\arguments{
\item{light_availability_spline_tol}{Error tolerance of adpative spline method. Deafult is 1e-4.}

\item{light_availability_spline_nbase}{Parameter used in adaptive spline method. Default is 17.}

\item{light_availability_spline_max_depth}{Parameter used in adaptive spline method. Default is 16.}

\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}

\item{soil_number_of_depths}{the number of soil layers}

\item{soil_initial_state}{the initial state of the soil layers}

\item{rainfall}{constant function value for rainfall driver, y = rainfall}
}
\description{
create a model Environment object
}
