% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R
\name{FF16_make_environment}
\alias{FF16_make_environment}
\title{create FF16_environment object}
\usage{
FF16_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)
}
\arguments{
\item{light_availability_spline_tol}{}

\item{light_availability_spline_nbase}{}

\item{light_availability_spline_max_depth}{}

\item{light_availability_spline_rescale_usually}{}
}
\description{
create FF16_environment object
}
