% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R
\name{FF16_make_environment}
\alias{FF16_make_environment}
\title{create FF16_environment object}
\usage{
FF16_make_environment(
  canopy_light_tol = 1e-04,
  canopy_light_nbase = 17,
  canopy_light_max_depth = 16,
  canopy_rescale_usually = TRUE
)
}
\arguments{
\item{canopy_light_tol}{}

\item{canopy_light_nbase}{}

\item{canopy_light_max_depth}{}

\item{canopy_rescale_usually}{}
}
\description{
create FF16_environment object
}
