% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R, r/k93.R, r/tf24.R
\name{FF16_Parameters}
\alias{FF16_Parameters}
\alias{K93_Parameters}
\alias{TF24_Parameters}
\title{Setup an a model system with default or specified parameters}
\usage{
FF16_Parameters(...)

K93_Parameters()

TF24_Parameters(...)
}
\arguments{
\item{...}{Arguments to be passed to the model constructor. These include

  *`patch_area`: Area of idnividfual patch. Only relevant for stochastic model. Default is 1.0m2.
  *`max_patch_lifetime`: The maximum time in years we want to simulate
  *`strategies`: A list of stratgies to simulate. The default is an empty list.
  *`strategy_default`: Values for the default startegy. The default values are those specified in the C++ code for the model.
  *`node_schedule_times_default`: Default vector of times at which to introduce nodes. The default is chosen to have close spacing at the start of the simulation.
  *`node_schedule_times`: A list with each element containing the vector of times we want to introduce nodes for each strategy. The default is an empty list.
  *`ode_times`: A vector of patch ages we want the ode solver to stop at}
}
\description{
Setup an a model system with default or specified parameters. 
This function enables you initialize a model system. Use the model name to start different models.
}
\examples{
p1 <- FF16_Parameters()
p2 <- FF16_Parameters(max_patch_lifetime = 10.0, patch_area = 1.0, strategies = list(FF16_Strategy()), strategy_default = FF16_Strategy(), node_schedule_times_default = node_schedule_times_default(10.0), node_schedule_times = list(node_schedule_times_default(10.0)), ode_times = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
p1 <- TF24_Parameters()
p2 <- TF24_Parameters(max_patch_lifetime = 10.0)
}
