% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16.R
\name{FF16_solve_max_size_growth_rate_at_height}
\alias{FF16_solve_max_size_growth_rate_at_height}
\title{Solves the maximum growth rate at a given height within the interval of the bounds of a given trait}
\usage{
FF16_solve_max_size_growth_rate_at_height(
  bounds,
  log_scale = TRUE,
  tol = 0.001,
  height = 1,
  type = "FF16",
  params = scm_base_parameters(type),
  env = make_environment(type = type),
  outcome = "height",
  use_optim = FALSE,
  hyperpars = hyperpar(type),
  use_default_strategy = TRUE
)
}
\arguments{
\item{outcome}{}
}
\description{
Solves the maximum growth rate at a given height within the interval of the bounds of a given trait
}
\author{
Isaac Towers, Daniel Falster and Andrew O'Reilly-Nugent
}
