% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R, r/ff16w.R, r/k93.R, r/tf24.R
\name{FF16_test_environment}
\alias{FF16_test_environment}
\alias{FF16w_test_environment}
\alias{K93_test_environment}
\alias{TF24_test_environment}
\title{Create a test environment for FF16 startegy}
\usage{
FF16_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)

FF16w_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)

K93_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)

TF24_test_environment(height, n = 101, light_env = NULL, n_strategies = 1)
}
\arguments{
\item{height}{top height of environment object}

\item{n}{number of points}

\item{light_env}{function for light environment in test object}

\item{n_strategies}{number of strategies for test environment}
}
\description{
This makes a pretend light environment over the plant height,
slightly concave up, whatever.
}
\examples{
environment <- plant:::FF16_test_environment(10)
environment <- plant:::FF16w_test_environment(10)
environment <- plant:::K93_test_environment(10)
environment <- plant:::TF24_test_environment(10)
}
