% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R, r/ff16r.R
\name{FF16r_Environment}
\alias{FF16r_Environment}
\alias{FF16r_make_environment}
\alias{FF16r_fixed_environment}
\title{FF16r_Environment object}
\usage{
FF16r_Environment(disturbance_mean_interval, seed_rain, k_I, control)

FF16r_make_environment(p)

FF16r_fixed_environment(e = 1, p = FF16r_Parameters(), height_max = 150)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{k_I}{light extinction coefficient}

\item{control}{A Control object}

\item{p}{A Parameters object}

\item{height_max}{= 150.0 maximum possible height in environment}

\item{e=1.0}{Value of environment}
}
\description{
FF16r_Environment object

Construct a fixed environment for FF16r strategy
}
