% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff16r.R
\name{make_FF16r_hyperpar}
\alias{FF16r_hyperpar}
\alias{make_FF16r_hyperpar}
\title{Hyperparameters for FF16r physiological model}
\usage{
make_FF16r_hyperpar(lma_0 = 0.1978791, B_kl1 = 0.4565855, B_kl2 = 1.71,
  rho_0 = 608, B_dI1 = 0.01, B_dI2 = 0, B_ks1 = 0.2, B_ks2 = 0,
  B_rs1 = 4012, B_rb1 = 2 * 4012, B_f1 = 3, narea_0 = 0.00187,
  B_lf1 = 5120.738 * 24 * 3600/1e+06, B_lf2 = 0.5, B_lf3 = 0.04,
  B_lf4 = 21000, k_I = 0.5, latitude = 0)

FF16r_hyperpar(m, s, filter = TRUE)
}
\arguments{
\item{lma_0}{Central (mean) value for leaf mass per area}

\item{B_kl1}{Rate of leaf turnover at phi_0}

\item{B_kl2}{Scaling slope for phi in leaf turnover}

\item{rho_0}{Central (mean) value for wood density}

\item{B_dI1}{Rate of instantaneous mortality at rho_0}

\item{B_dI2}{Scaling slope for wood density in intrinsic mortality}

\item{B_ks1}{Rate of sapwood turnover at rho_0}

\item{B_ks2}{Scaling slope for rho in sapwood turnover}

\item{B_rs1}{CO_2 respiration per unit sapwood volume}

\item{B_rb1}{CO_2 respiration per unit sapwood volume}

\item{B_f1}{Cost of seed accessories per unit seed mass}

\item{narea_0}{central (mean) value for nitrogen per leaf area}

\item{B_lf1}{Potential CO_2 photosynthesis per unit narea}

\item{B_lf2}{Curvature of leaf photosynthetic light response curve}

\item{B_lf3}{Quantum yield of leaf photosynthetic light response curve}

\item{B_lf4}{CO_2 respiration per unit leaf nitrogen}

\item{k_I}{light extinction coefficient}

\item{latitude}{degrees from equator (0-90), used in solar model}
}
\description{
Hyperparameters for FF16r physiological model

Hyperparameters for FF16r physiological model
}

