% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{make_ff_parameters}
\alias{ff_parameters}
\alias{make_ff_parameters}
\title{Hyperparameters for tree}
\usage{
make_ff_parameters(B4 = 1.71)

ff_parameters(m)
}
\arguments{
\item{B4}{Slope of lma / leaf turnover log-log relationship}

\item{m}{A trait matrix}
}
\description{
Hyperparameters for tree
}

