% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{make_FFW16_hyperpar}
\alias{FFW16_hyperpar}
\alias{FFdev_hyperpar}
\alias{hyperpar}
\alias{make_FFW16_hyperpar}
\alias{make_hyperpar}
\title{Hyperparameters for plant}
\usage{
make_FFW16_hyperpar(lma_0 = 0.1978791, B_kl1 = 0.4565855, B_kl2 = 1.71,
  rho_0 = 608, B_dI1 = 0.01, B_dI2 = 0, B_ks1 = 0.2, B_ks2 = 0,
  B_rs1 = 4012, B_rb1 = 2 * 4012, B_f1 = 3, narea_0 = 0.00187,
  B_lf1 = 5120.738 * 24 * 3600/1e+06, B_lf2 = 0.5, B_lf3 = 0.04,
  B_lf4 = 21000, k_I = 0.5, latitude = 0)

FFW16_hyperpar(m, s, filter = TRUE)

FFdev_hyperpar(m, s, filter = TRUE)

make_hyperpar(type)

hyperpar(type)
}
\arguments{
\item{lma_0}{LMA value...}

\item{B_kl1}{...}

\item{B_kl2}{...}

\item{rho_0}{...}

\item{B_dI1}{...}

\item{B_dI2}{...}

\item{B_ks1}{...}

\item{B_ks2}{...}

\item{B_rs1}{...}

\item{B_rb1}{...}

\item{B_f1}{...}

\item{narea_0}{...}

\item{B_lf1}{...}

\item{B_lf2}{...}

\item{B_lf3}{...}

\item{B_lf4}{...}

\item{k_I}{...}

\item{latitude}{...}

\item{m}{A trait matrix}

\item{s}{A default strategy}

\item{filter}{Logical, indicating if generated parameters that are
the same as the default should be removed.}

\item{type}{Either \code{"FFW16"} or \code{"FFdev"}.}
}
\description{
Hyperparameters for plant
}

