% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_singularity_1D}
\alias{find_singularity_1D}
\title{Find evolutionary attractor for single species and trait.}
\usage{
find_singularity_1D(trait, interval, p, tol = 0.001, ...)
}
\arguments{
\item{trait}{name of trait}

\item{interval}{a vector containing the end-points of the
interval to be searched for the root.}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{...}{set verbpse=TRUE for verbose output, seed_rain=value
gives starting values when solving for demographic equilibrium}

\item{tol=1e-4}{the desired accuracy (convergence tolerance).}
}
\value{
A species object, with trait, seed rain and cohort schedule
information.
}
\description{
Find evolutionary attractor for single species and trait.
This is point at which selection gradient equals zero.
Currently solved using \code{uniroot}
}
\author{
Daniel Falster
}

