% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{grow_individual_to_size}
\alias{grow_individual_to_size}
\alias{grow_individual_to_height}
\title{Grow individual to given size}
\usage{
grow_individual_to_size(
  individual,
  sizes,
  size_name,
  env,
  time_max = Inf,
  warn = TRUE,
  filter = FALSE
)

grow_individual_to_height(individual, heights, env, ...)
}
\arguments{
\item{individual}{An \code{Individual} object.}

\item{sizes}{A vector of sizes to grow the plant to (increasing in
size).}

\item{size_name}{The name of the size variable within
\code{individual$rates} (e.g., height).}

\item{env}{An \code{Environment} object.}

\item{time_max}{Time to run the ODE out for -- only exists to
prevent an infinite loop (say, on an unreachable size).}

\item{warn}{Warn if requesting a plant that is too large?}

\item{filter}{Filter individuals that are too large?}

\item{heights}{Heights (when using \code{grow_individual_to_height})}

\item{...}{Additional parameters passed to
\code{grow_individual_to_size}.}
}
\value{
A list with elements \code{time} (the time that a given
size was reached), \code{state} (the \emph{ode state} at these
times, as a matrix) and \code{plant} a list of individuals grown to the
appropriate size.  Note that if only a single size is given,
a list of length 1 is returned.
}
\description{
Grow an individual up to particular sizes.
}
