% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\name{grow_individual_to_time}
\alias{grow_individual_to_time}
\title{Grow a plant}
\usage{
grow_individual_to_time(individual, times, env)
}
\arguments{
\item{individual}{An \code{Individual} object}

\item{times}{A vector of times}

\item{env}{An \code{Environment} object}
}
\description{
Grow a plant up for particular time lengths
}
