% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grow_plant_to_size}
\alias{grow_plant_to_size}
\title{Grow plant to given size}
\usage{
grow_plant_to_size(plant, sizes, env, size_name, time_max = Inf)
}
\arguments{
\item{plant}{A \code{Plant} object.}

\item{env}{An \code{Environment} object.}

\item{size_name}{The name of the size variable within
\code{Plant$vars_phys} (e.g., height).}

\item{time_max}{Time to run the ODE out for -- only exists to
prevent an infinite loop (say, on an unreachable size).}
}
\value{
A list with elements \code{time} (the time that a given
size was reached), \code{state} (the \emph{ode state} at these
times, as a matrix) and \code{plant} a list of plants grown to the
appropriate size.
}
\description{
Grow a plant up to particular sizes.
}

