% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{make_hyperpar}
\alias{make_hyperpar}
\alias{hyperpar}
\title{Hyperparameters for FF16 physiological model}
\usage{
make_hyperpar(type)

hyperpar(type)
}
\arguments{
\item{type}{Either \code{"FF16"} or \code{"FF16r"}.}

\item{type}{Either \code{"FF16"} or \code{"FF16r"}.}
}
\description{
Set a suitable hyperparameter function for chosen physiological model
}
