% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/strategy_support.R
\name{make_hyperpar}
\alias{make_hyperpar}
\alias{hyperpar}
\title{Hyperparameters for FF16 physiological model}
\usage{
make_hyperpar(type)

hyperpar(type)
}
\arguments{
\item{type}{Any strategy name as a string, e.g.: \code{"FF16"}.}
}
\description{
Set a suitable hyperparameter function for chosen physiological model
}
