% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{Internals}
\alias{Internals}
\title{Internals class holding vectors of states, thier associated rates and auxiliary (aux) state 
which is calculated from the state through running compute_rates}
\usage{
Internals(s_size, a_size)
}
\description{
Internals class holding vectors of states, thier associated rates and auxiliary (aux) state 
which is calculated from the state through running compute_rates
}
