% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_outputs.R
\name{interpolate_to_heights}
\alias{interpolate_to_heights}
\title{Interpolate data on size distributions for each species to specific heights at every time point}
\usage{
interpolate_to_heights(
  tidy_species_data,
  heights,
  method = "natural",
  min_log_density = -100
)
}
\arguments{
\item{tidy_species_data}{output of either `tidy_patch` or `tidy_species`}

\item{heights}{heights to interpolate to}

\item{method}{Method for interpolation. For more info see help on stats::spline}
}
\value{
Returns a tibble of similar format to input, but with all outputs interpolated to specified hieghts.
}
\description{
Interpolate data on size distributions for each species to specific heights at every time point
}
