% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_patch.R
\name{interpolate_to_heights}
\alias{interpolate_to_heights}
\title{Interpolate tidyspecies data to specific heights at every time point}
\usage{
interpolate_to_heights(tidy_species_data, heights, method = "natural")
}
\arguments{
\item{tidy_species_data}{???}

\item{heights}{heights to interpolate to}

\item{method}{Method for interpolation. For more info see help on stats::spline}
}
\value{
??
}
\description{
Interpolate tidyspecies data to specific heights at every time point
}
