% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_outputs.R
\name{interpolate_to_times}
\alias{interpolate_to_times}
\title{Interpolate data on size distributions for each species to specific timer points, using specified interpolation method}
\usage{
interpolate_to_times(tidy_species_data, times, method = "natural")
}
\arguments{
\item{tidy_species_data}{output of either `tidy_patch` or `tidy_species`}

\item{times}{times to interpolate to}

\item{method}{Method for interpolation. For more info see help on stats::spline}
}
\value{
Returns a tibble of similar format to input, but with all outputs interpolated to specified hieghts.
}
\description{
Interpolate data on size distributions for each species to specific timer points, using specified interpolation method
}
