% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/k93.R
\name{K93_Environment}
\alias{K93_Environment}
\alias{K93_fixed_environment}
\title{K93_Environment object}
\usage{
K93_Environment()

K93_fixed_environment(e = 1, height_max = 300)
}
\arguments{
\item{e}{Value of environment (default=1.0)}

\item{height_max}{= 300.0 maximum possible height in environment}
}
\description{
K93_Environment object

Construct a fixed environment for K93 strategy
}
