% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R, r/k93.R
\name{K93_Environment}
\alias{K93_Environment}
\alias{K93_make_environment}
\alias{K93_fixed_environment}
\title{K93_Environment object}
\usage{
K93_Environment(disturbance_mean_interval, seed_rain, k_I, control)

K93_make_environment(p)

K93_fixed_environment(e = 1, p = K93_Parameters(), height_max = 300)
}
\arguments{
\item{disturbance_mean_interval}{mean disturbance interval in years}

\item{seed_rain}{vector of seed rains}

\item{k_I}{light extinction coefficient}

\item{control}{A Control object}

\item{p}{A Parameters object}

\item{height_max}{= 150.0 maximum possible height in environment}

\item{e=1.0}{Value of environment}
}
\description{
K93_Environment object

Construct a fixed environment for K93 strategy
}
