% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/k93.R
\name{make_K93_hyperpar}
\alias{make_K93_hyperpar}
\alias{K93_hyperpar}
\title{Hyperparameters for K93 physiological model}
\usage{
make_K93_hyperpar(
  b_0 = 0.059,
  b_1 = 0.012,
  b_2 = 0.00041,
  c_0 = 0.008,
  c_1 = 0.00044,
  d_0 = 0.00073,
  d_1 = 0.044
)

K93_hyperpar(m, s, filter = TRUE)
}
\arguments{
\item{m}{A matrix of trait values, as returned by \code{trait_matrix}}

\item{s}{A strategy object}

\item{filter}{A flag indicating whether to filter columns. If TRUE, any numbers 
that are within eps of the default strategy are not replaced.}
}
\description{
Hyperparameters for K93 physiological model

Hyperparameter function for K93 physiological model
}
