% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k93.R
\name{K93_make_environment}
\alias{K93_make_environment}
\title{create K93_environment object}
\usage{
K93_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)
}
\arguments{
\item{light_availability_spline_tol}{}

\item{light_availability_spline_rescale_usually}{}
}
\description{
create K93_environment object
}
