% Generated by roxygen2 (4.0.2): do not edit by hand
\name{landscape}
\alias{landscape}
\alias{landscape_empty}
\title{Fitness Landscape}
\usage{
landscape(trait, values, p, schedule = NULL, raw_seed_rain = FALSE)

landscape_empty(trait, values, p, schedule = NULL, raw_seed_rain = FALSE)
}
\arguments{
\item{trait}{Name of the single trait to change}

\item{values}{Vector of trait values to compute mutant fitness}

\item{p}{Parameters object.  Needs to contain residents with their
incoming seed rain.}

\item{schedule}{Schedule of times for the residents.}

\item{raw_seed_rain}{Should the raw (per capita) seed rain values
be returned, rather than converting to our measure of fitness (log
per capita seed rain, being the instantaneous rate of population
increase).}
}
\value{
Vector with the output seed rain.  Mutants have an
arbitrary seed rain of zero, so this is the rate of seed
production per capita.
}
\description{
Construct a fitness landscape.  Currently only works in one
dimension, and with all other parameters set to their defaults,
and at predefined values only rather than adaptively (i.e., not
actually useful yet).
}
\details{
Will change...
}
\author{
Rich FitzJohn
}

