% Generated by roxygen2 (4.0.2): do not edit by hand
\name{last}
\alias{first}
\alias{first<-}
\alias{last}
\alias{last<-}
\title{Get Last Element}
\usage{
last(x)

last(x) <- value

first(x)

first(x) <- value
}
\arguments{
\item{x}{An object that can be subset with \code{[[}}
}
\description{
Get last element from an object

##'
}
\author{
Rich FitzJohn
}

