% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\name{lcp_whole_plant}
\alias{lcp_whole_plant}
\title{Whole plant light compensation point}
\usage{
lcp_whole_plant(p, ...)
}
\arguments{
\item{p}{A \code{PlantPlus}, with strategy, height, etc set.}

\item{...}{Additional arguments that are ignored}
}
\description{
Compute the whole plant light compensation point for a single
plant.
}
\author{
Rich FitzJohn
}
