% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Leaf}
\alias{Leaf}
\title{Leaf model}
\usage{
Leaf(
  vcmax_25,
  c,
  b,
  psi_crit,
  beta2,
  jmax_25,
  hk_s,
  a,
  curv_fact_elec_trans,
  curv_fact_colim,
  GSS_tol_abs,
  vulnerability_curve_ncontrol,
  ci_abs_tol,
  ci_niter,
  g0,
  g1,
  g1_TF24
)
}
\arguments{
\item{vcmax_25}{Maximum Carboxylation Rate (umol m^-2 s^-1)}

\item{c}{shape parameter of hydraulic vulnerability curve (unitless)}

\item{b}{sensitivity parameter of hydraulic vulnerability curve (MPa)}

\item{psi_crit}{critical psi stem (-MPa )}

\item{beta2}{exponent for effect of hydraulic risk (unitless)}

\item{jmax_25}{maximum rate of electron transport (umol m^-2 s^-1)}

\item{hk_s}{maximum hydraulic-dependent sapwood turnover rate yr ^ -1}

\item{a}{quantum yield of photosynthetic electron transport (mol mol^-1)}

\item{curv_fact_elec_trans}{curvature factor for the light response curve (unitless)}

\item{curv_fact_colim}{curvature factor for the colimited photosythnthesis equatiom}

\item{GSS_tol_abs}{tolereance parameter for golden section search}

\item{vulnerability_curve_ncontrol}{number of points for pre-calculating vulnerability curve}

\item{ci_abs_tol}{absolute tolerance value for root-solving ci}

\item{ci_niter}{maximum number of iterations for root-solving ci}

\item{g0}{g0 parameter in the medlyn model umol m^-2 s^-1}

\item{g1}{parameter in the medlyn model umol kPa^0.5}

\item{g1_TF24}{cost parameter for TF24 profit model umol m^-2 s^-1}
}
\description{
Leaf model
}
