% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Leaf}
\alias{Leaf}
\title{Leaf model}
\usage{
Leaf(
  vcmax_25,
  c,
  b,
  psi_crit,
  root_c,
  root_b,
  root_psi_crit,
  beta2,
  jmax_25,
  hk_s,
  a,
  curv_fact_elec_trans,
  curv_fact_colim,
  GSS_tol_abs,
  vulnerability_curve_ncontrol,
  ci_abs_tol,
  ci_niter,
  g1_TF24,
  beta_R_H,
  beta_R_V
)
}
\arguments{
\item{vcmax_25}{Maximum Carboxylation Rate (umol m^-2 s^-1)}

\item{c}{shape parameter of hydraulic vulnerability curve (unitless)}

\item{b}{sensitivity parameter of hydraulic vulnerability curve (MPa)}

\item{psi_crit}{critical psi stem (-MPa )}

\item{root_c}{shape parameter of hydraulic vulnerability curve for roots(unitless)}

\item{root_b}{sensitivity parameter of hydraulic vulnerability curve for roots (MPa)}

\item{root_psi_crit}{critical psi stem for roots (-MPa )}

\item{beta2}{exponent for effect of hydraulic risk (unitless)}

\item{jmax_25}{maximum rate of electron transport (umol m^-2 s^-1)}

\item{hk_s}{maximum hydraulic-dependent sapwood turnover rate yr ^ -1}

\item{a}{quantum yield of photosynthetic electron transport (mol mol^-1)}

\item{curv_fact_elec_trans}{curvature factor for the light response curve (unitless)}

\item{curv_fact_colim}{curvature factor for the colimited photosythnthesis equatiom}

\item{GSS_tol_abs}{tolereance parameter for golden section search}

\item{vulnerability_curve_ncontrol}{number of points for pre-calculating vulnerability curve}

\item{ci_abs_tol}{absolute tolerance value for root-solving ci}

\item{ci_niter}{maximum number of iterations for root-solving ci}

\item{g1_TF24}{cost parameter for TF24 profit model umol m^-2 s^-1}

\item{beta_R_H}{proportionality constant between minimum horizontal (intralayer) root hydraulic resistance and C_r^-1 (MPa s mol C / mol H2O)}

\item{beta_R_V}{proportionality constant between minimum vertical (interlayer) root hydraulic resistance and dz^2/C_r (MPa mol C s / mol H2O / m^2)}
}
\description{
Leaf model
}
