% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{make_patch}
\alias{ebt_patch}
\alias{ebt_state}
\alias{make_patch}
\title{Reconstruct a patch}
\usage{
make_patch(state, p)

ebt_state(i, x)

ebt_patch(i, x)
}
\arguments{
\item{state}{State object created by \code{ebt_state}}

\item{p}{Parameters object}

\item{i}{Index to extract from \code{x}}

\item{x}{Result of running \code{\link{run_ebt_collect}}}
}
\description{
Functions for reconstructing a Patch from an EBT
}

