% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{make_patch}
\alias{make_patch}
\alias{scm_patch}
\title{Reconstruct a patch}
\usage{
make_patch(
  state,
  p,
  env = make_environment(parameters = p),
  ctrl = scm_base_control()
)

scm_patch(i, x)
}
\arguments{
\item{state}{State object created by \code{scm_state}}

\item{p}{Parameters object}

\item{env}{Environment object (defaults to FF16_Environment)}

\item{ctrl}{Control object}
}
\description{
Functions for reconstructing a Patch from an SCM
}
