% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/scm_support.R
\name{make_patch}
\alias{make_patch}
\alias{scm_state}
\alias{scm_patch}
\title{Reconstruct a patch}
\usage{
make_patch(state, p)

scm_state(i, x)

scm_patch(i, x)
}
\arguments{
\item{state}{State object created by \code{scm_state}}

\item{p}{Parameters object}

\item{i}{Index to extract from \code{x}}

\item{x}{Result of running \code{\link{run_scm_collect}}}
}
\description{
Functions for reconstructing a Patch from an SCM
}
