% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_plant.R
\name{make_reference_plant}
\alias{make_reference_plant}
\title{Reference Growth Model}
\usage{
make_reference_plant(type = "FF16")
}
\arguments{
\item{type}{name of model to make a reference plant for (only FF16
supported at present).}
}
\value{
An object with a number of methods (accessed with
\code{$}).  These will hopefully be documented at some point.
}
\description{
Create an instance of reference version for specified growth model
}
\author{
Daniel S. Falster (extra code and porting by Rich
FitzJohn)
}
