% Generated by roxygen2 (4.0.1): do not edit by hand
\name{max_fitness}
\alias{max_fitness}
\title{Find point of maximum fitness within some range.}
\usage{
max_fitness(trait, p, bounds = NULL, log_scale = TRUE)
}
\arguments{
\item{trait}{Name of the trait (e.g., \code{"lma"})}

\item{bounds}{2D vector specifing range within which to search}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{log_scale}{Is the parameter naturally on a log scale?  If
so, this will greatly speed things up.}
}
\description{
Find point of maximum fitness in empty landscape
within specified range.
}
\author{
Daniel Falster, Rich FitzJohn
}

