% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/fitness.R
\name{max_growth_rate}
\alias{max_growth_rate}
\title{Compute Max Growth Rate}
\usage{
max_growth_rate(trait_matrix, p)
}
\arguments{
\item{trait_matrix}{A matrix of traits to compute maximum growth
rate at (see \code{\link{trait_matrix}})}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}
}
\description{
Compute max growth rate of a given set of values for a trait.
This is the log of per-capita seed production (i.e., fitness).
}
\details{
Only works in one dimension
}
\author{
Rich FitzJohn
}
