% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{NodeSchedule}
\alias{NodeSchedule}
\title{Schedule of node introduction times}
\usage{
NodeSchedule(n_species)
}
\arguments{
\item{n_species}{number of species}
}
\description{
Schedule of node introduction times
}
