% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/RcppR6.R
\name{OdeControl}
\alias{OdeControl}
\title{ODE Control parameters}
\usage{
OdeControl(..., values = list(...))
}
\arguments{
\item{..., values}{Values to initialise the struct with (either as
variadic arguments, or as a list, but not both).}
}
\description{
Control parameters for the ode system
}
