% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{optimise_individual_rate_at_size_by_trait}
\alias{optimise_individual_rate_at_size_by_trait}
\alias{optimise_individual_rate_at_height_by_trait}
\title{The function `optimise_individual_rate_at_height_by_trait` and `optimise_individual_rate_at_size_by_trait` solve for the maximum of
some rate (e.g. growth rate) at a specified height within
the interval of the bounds of a given trait}
\usage{
optimise_individual_rate_at_size_by_trait(
  type = "FF16",
  bounds,
  log_scale = TRUE,
  tol = 0.001,
  size = 1,
  size_name = "height",
  rate = size_name,
  params = scm_base_parameters(type),
  env = make_environment(type),
  hyperpars = hyperpar(type),
  set_state_directly = FALSE
)

optimise_individual_rate_at_height_by_trait(..., height = 1)
}
\arguments{
\item{hyperpars}{}
}
\description{
The function `optimise_individual_rate_at_height_by_trait` and `optimise_individual_rate_at_size_by_trait` solve for the maximum of
some rate (e.g. growth rate) at a specified height within
the interval of the bounds of a given trait
}
\author{
Isaac Towers, Daniel Falster and Andrew O'Reilly-Nugent
}
