% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Plant}
\alias{Plant}
\title{Plant object}
\usage{
Plant(T)
}
\arguments{
\item{T}{String containing class of plant to create; currently
must be \code{"FFW16"}.  See \code{\link{FFW16_Plant}} for more
details.}
}
\description{
Plant object
}

