% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Plant}
\alias{Plant}
\title{Plant object}
\usage{
Plant(T)
}
\arguments{
\item{T}{String containing class of plant to create; currently
must be \code{"FF16"}.  See \code{\link{FF16_Plant}} for more
details.}
}
\description{
Plant object
}
