% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{resource_compensation_point}
\alias{resource_compensation_point}
\title{Whole plant light compensation point}
\usage{
resource_compensation_point(p, ...)
}
\arguments{
\item{p}{A \code{PlantPlus}, with strategy, height, etc set.}

\item{...}{Additional arguments that are ignored}
}
\description{
Compute the whole plant light compensation point for a single
plant.
}
\author{
Rich FitzJohn
}
