% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run.reference}
\alias{evolve.is.installed}
\alias{install.evolve}
\alias{run.reference}
\alias{uninstall.evolve}
\title{Run Reference Model}
\usage{
run.reference(path, p = NULL, verbose = TRUE, evolve = NULL)

install.evolve(reinstall = FALSE, verbose = FALSE, url = NULL)

uninstall.evolve()

evolve.is.installed()
}
\arguments{
\item{path}{Path to directory containing parameter files (see
\code{\link{load.reference.output}} and
\code{\link{load.reference.parameters}}}

\item{p}{Optional \code{\link{Parameters}} object that will be
written to \code{path} before running model (equivalent to "run
model on these parameters in this directory")}

\item{verbose}{Logical indicating if the model output should be
printed (TRUE by default).}

\item{evolve}{Path to the "evolve" program.  If not specified,
"evolve" is assumed to be installed within tree by
\code{install.evolve}}
}
\description{
Run the reference model, falster-traitdiversity
}
\details{
This function is known to work with version d649534 of
falster-traitdiversity (and will not work with any earlier
version).
}
\author{
Rich FitzJohn
}

