% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_ebt}
\alias{run_ebt}
\title{Run EBT}
\usage{
run_ebt(p, use_ode_times = FALSE)
}
\arguments{
\item{p}{Parameters object}

\item{use_ode_times}{Should ODE times be used?}
}
\value{
A \code{EBT} object.
}
\description{
Run the EBT, returning the EBT object for interrogation
}
\details{
This is the simplest way of using the EBT, probably.
}
\author{
Rich FitzJohn
}

