% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ebt_support.R
\name{run_ebt_collect}
\alias{run_ebt_collect}
\title{Run the EBT, Collecting Output}
\usage{
run_ebt_collect(p)
}
\arguments{
\item{p}{A Parameters object}

\item{sched}{A CohortSchedule Object}
}
\description{
Run the EBT model, given a Parameters and CohortSchedule
}
\details{
This is mostly a simple wrapper around some of the EBT functions.
Not sure if this is how we will generally want to do this.
Consider this function liable to change.
}
\author{
Rich FitzJohn
}

