% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{run_scm}
\alias{run_scm}
\title{Run SCM}
\usage{
run_scm(
  p,
  env = NULL,
  ctrl = scm_base_control(),
  use_ode_times = FALSE,
  collect = FALSE
)
}
\arguments{
\item{p}{Parameters object}

\item{env}{Environment object (defaults to FF16_Environment)}

\item{ctrl}{Control object}

\item{use_ode_times}{Should ODE times be used?}
}
\value{
A \code{SCM} object.
}
\description{
Run the SCM, returning the SCM object for interrogation
}
\details{
This is the simplest way of using the SCM, probably.
}
\author{
Rich FitzJohn
}
