% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{run_scm_collect}
\alias{run_scm_collect}
\title{Run the SCM, Collecting Output}
\usage{
run_scm_collect(
  p,
  env = Environment(parameters = p),
  ctrl = scm_base_control(),
  collect_auxiliary_variables = FALSE
)
}
\arguments{
\item{p}{A \code{Parameters} object}

\item{env}{Environment object (defaults to FF16_Environment)}

\item{ctrl}{Control object}

\item{collect_auxiliary_variables}{Return additional strategy variables (eg
competition_effect)}
}
\description{
Run the SCM model, given a Parameters and NodeSchedule
}
\details{
This is mostly a simple wrapper around some of the SCM functions.
Not sure if this is how we will generally want to do this.
Consider this function liable to change.
}
\author{
Rich FitzJohn
}
