% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/scm_support.R
\name{run_scm_collect}
\alias{run_scm_collect}
\title{Run the SCM, Collecting Output}
\usage{
run_scm_collect(p, include_competition_effect = FALSE)
}
\arguments{
\item{p}{A \code{Parameters} object}

\item{include_competition_effect}{Include total leaf area (will change; see
issue #138)}
}
\description{
Run the SCM model, given a Parameters and CohortSchedule
}
\details{
This is mostly a simple wrapper around some of the SCM functions.
Not sure if this is how we will generally want to do this.
Consider this function liable to change.
}
\author{
Rich FitzJohn
}
