% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/scm_support.R
\name{scm_base_parameters}
\alias{scm_base_parameters}
\title{Basic default parameters for a given strategy}
\usage{
scm_base_parameters(type = NA, env = environment_type(type))
}
\arguments{
\item{type}{Any strategy name as a string, e.g.: \code{"FF16"}.}

\item{env}{And environment object}
}
\description{
Basic default settings for a given strategy, environment only really
used for templating initially and will be overloaded later by passing
an environment to the SCM API (suggesting perhaps the template could be
removed).
}
\author{
Rich FitzJohn
}
