% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_support.R
\name{scm_base_parameters}
\alias{scm_base_parameters}
\title{Sensible, fast (ish) SCM parameters}
\usage{
scm_base_parameters(type = "FF16")
}
\arguments{
\item{type}{Name of model (defaults to FF16 but FF16r also valid)}
}
\description{
Hopefully sensible set of parameters for use with the SCM.  Turns
accuracy down a bunch, makes it noisy, sets up the
hyperparameterisation that we most often use.
}
\author{
Rich FitzJohn
}
