% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq_log}
\alias{seq_log}
\alias{seq_log_range}
\title{Sequence in Log Space}
\usage{
seq_log(from, to, length.out)

seq_log_range(r, n)
}
\arguments{
\item{from}{Starting point}

\item{to}{Ending point}

\item{length.out}{Number of points to generate}

\item{r}{range (i.e., c(from, to)}
}
\description{
Sequence in log space
}
\details{
Unlike the billions of options for \code{seq}, only
\code{length.out} is supported here, and both \code{from} and
\code{to} must be provided.
}
\author{
Rich FitzJohn
}

