% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_log}
\alias{seq_log}
\alias{seq_log_range}
\alias{seq_range}
\title{Sequence in log space}
\usage{
seq_log(from, to, length.out)

seq_log_range(r, length.out)

seq_range(r, length.out)
}
\arguments{
\item{from}{Starting point}

\item{to}{Ending point}

\item{length.out}{Number of points to generate}

\item{r}{range (i.e., c(from, to)}
}
\description{
Sequence in log space
}
\details{
Unlike the billions of options for \code{seq}, only
\code{length.out} is supported here, and both \code{from} and
\code{to} must be provided.  For completeness, \code{seq_range}
generates a range in non-log space.
}
\author{
Rich FitzJohn
}
