% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_trait_space.R
\name{solve_max_fitness}
\alias{solve_max_fitness}
\title{Find point of maximum fitness within some range.}
\usage{
solve_max_fitness(bounds, params, log_scale = TRUE, tol = 0.001)
}
\arguments{
\item{bounds}{Two element vector specifying range within which to
search}

\item{params}{Parameters for fundamental_fitness function}

\item{log_scale}{Is the parameter naturally on a log scale?  If
so, this will greatly speed things up.}

\item{tol}{Tolerance used in the optimisation}
}
\description{
Find point of maximum fitness in empty fitness landscape within a
specified range.
}
\author{
Daniel Falster, Rich FitzJohn
}
