% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{strategy_list}
\alias{plant_list}
\alias{strategy_list}
\title{Create a list of Strategies}
\usage{
strategy_list(x, strategy = Strategy())

plant_list(x, strategy = Strategy())
}
\arguments{
\item{x}{Values for the trait.  This must be a \emph{matrix}, with
column names corresponding to entries in \code{Strategy} and rows
representing different values.}

\item{strategy}{Default strategy to modify}
}
\description{
Create a list of Strategies or Plants by varying a single trait.
}

