% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strategy_list}
\alias{plant_list}
\alias{strategy_list}
\title{Create a list of Strategies}
\usage{
strategy_list(x, trait_name, strategy = Strategy())

plant_list(x, trait_name, strategy = Strategy())
}
\arguments{
\item{x}{Values for the trait}

\item{trait_name}{Name of the trait to vary (may be any element of
Strategy)}

\item{strategy}{Default strategy to modify}
}
\description{
Create a list of Strategies or Plants by varying a single trait.
}

