% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{TF24_Environment}
\alias{TF24_Environment}
\title{Create an TF24_Environment object.}
\usage{
TF24_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths,
  delta_z,
  soil_moist_sat,
  K_sat,
  a_psi,
  n_psi,
  b_infil
)
}
\arguments{
\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}

\item{soil_number_of_depths}{Number of soil layers to include}

\item{delta_z}{Thickness of soil layers}

\item{soil_moist_sat}{Soil moisture at saturation}

\item{K_sat}{Saturated hydraulic conductivity}

\item{a_psi}{Coefficient for soil water potential}

\item{n_psi}{Exponent for soil water potential}

\item{b_infil}{Coefficient for infiltration}

\item{rainfall}{Rainfall amount}

\item{atm_vpd}{Atmospheric vapor pressure deficit}

\item{ca}{Atmospheric CO2 concentration}

\item{leaf_temp}{Leaf temperature}

\item{atm_o2_kpa}{Atmospheric O2 partial pressure}

\item{atm_kpa}{Atmospheric pressure}
}
\description{
This environment is used for the TF24 strategy, which includes soil moisture and hydraulic
conductivity parameters.
}
