% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{TF24_Environment}
\alias{TF24_Environment}
\title{Create an TF24_Environment object.}
\usage{
TF24_Environment()
}
\description{
This environment is used for the TF24 strategy, which includes soil moisture and hydraulic
conductivity parameters.
}
