% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R, R/tf24.R
\name{TF24_Environment}
\alias{TF24_Environment}
\alias{TF24_fixed_environment}
\title{TF24_Environment object}
\usage{
TF24_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths
)

TF24_fixed_environment(e = 1, height_max = 150)
}
\arguments{
\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}

\item{soil_number_of_depths}{Number of soil layers to include}

\item{e}{Value of environment (deafult  = 1.0)}

\item{height_max}{= 150.0 maximum possible height in environment}
}
\description{
TF24_Environment object

Construct a fixed environment for TF24 strategy
}
