% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_make_environment}
\alias{TF24_make_environment}
\title{create TF24_environment object}
\usage{
TF24_make_environment(
  light_availability_spline_tol = 1e-04,
  light_availability_spline_nbase = 17,
  light_availability_spline_max_depth = 16,
  light_availability_spline_rescale_usually = TRUE
)
}
\arguments{
\item{light_availability_spline_tol}{}

\item{light_availability_spline_nbase}{}

\item{light_availability_spline_max_depth}{}

\item{light_availability_spline_rescale_usually}{}
}
\description{
create TF24_environment object
}
