% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_patch.R
\name{tidy_env}
\alias{tidy_env}
\title{Turn `env` component of scm output into a tidy data object}
\usage{
tidy_env(env)
}
\arguments{
\item{data}{a list, the `env` component of scm output.}
}
\value{
a tibble describing the environment in a patch
}
\description{
Turn `env` component of scm output into a tidy data object
}
