% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_outputs.R
\name{tidy_species}
\alias{tidy_species}
\alias{tidy_env}
\alias{tidy_patch}
\title{Turn `species` component of plant solver output into a tidy data object}
\usage{
tidy_species(results)

tidy_env(results)

tidy_patch(results)
}
\arguments{
\item{results}{output of run_scm(..., collect = TRUE)}
}
\value{
a list, containing outputs of plant solver in tidy format
}
\description{
Turn `species` component of plant solver output into a tidy data object

Turn `env` component of solver output into a tidy data object

Turns output of plant solver into a tidy data object
}
