% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_patch.R
\name{tidy_patch}
\alias{tidy_patch}
\title{Turns scm output into a tidy data object}
\usage{
tidy_patch(results)
}
\arguments{
\item{data}{output of run_scm_collect}
}
\value{
a list
}
\description{
Turns scm output into a tidy data object
}
