% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_outputs.R
\name{tidy_patch}
\alias{tidy_patch}
\title{Turns output of plant solver into a tidy data object}
\usage{
tidy_patch(results)
}
\arguments{
\item{results}{output of run_scm_collect}
}
\value{
a list, containing outputs of plant solver in tidy format
}
\description{
Turns output of plant solver into a tidy data object
}
