% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_patch.R
\name{tidy_species}
\alias{tidy_species}
\title{Turn `species` component of scm output into a tidy data object}
\usage{
tidy_species(data)
}
\arguments{
\item{data}{a list, the `species` component of scm output.}
}
\value{
a tibble describing columns of the output for each cohort in each species
}
\description{
Turn `species` component of scm output into a tidy data object
}
