% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/tidy_outputs.R
\name{tidy_species}
\alias{tidy_species}
\title{Turn `species` component of plant solver output into a tidy data object}
\usage{
tidy_species(data)
}
\arguments{
\item{data}{a list, the `species` component of plant solver output.}
}
\value{
a tibble whose columns provide metrics on each breakpoint in species size distribution
}
\description{
Turn `species` component of plant solver output into a tidy data object
}
