% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/util_model.R
\name{trait_matrix}
\alias{trait_matrix}
\title{Create trait matrix}
\usage{
trait_matrix(x, trait_name)
}
\arguments{
\item{x}{Values}

\item{trait_name}{Name of a single trait}
}
\description{
Helper function to create trait matrices suitable for
\code{\link{strategy_list}}.
}
\author{
Rich FitzJohn
}
