% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate}
\alias{validate}
\title{Validate an object}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Additional arguments to be passed to methods}
}
\description{
Validate an object.  Currently only \code{Parameters} objects are
validated.
}
