% Generated by roxygen2 (4.0.0): do not edit by hand
\name{viable_fitness}
\alias{viable_fitness}
\title{Compute Region of Positive Fitnes}
\usage{
viable_fitness(trait, p, bounds = NULL, value = NULL, log_scale = TRUE,
  dx = 1)
}
\arguments{
\item{trait}{Name of the trait (e.g., \code{"lma"})}

\item{bounds}{2D vector specifing range within which to search}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{value}{Initial value - must have positive fitness itself!
If not given, then the value from the default strategy within
\code{p} is used.}

\item{log_scale}{Is the parameter naturally on a log scale?  If
so, this will greatly speed things up.}

\item{dx}{Amount to step the trait.  If \code{log_scale} is
\code{TRUE}, this is on a log scale.}
}
\description{
Compute region of positive fitness.  This will be the values where
fitness is approximately zero.
}
\author{
Rich FitzJohn
}

