% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_trait_space.R
\name{viable_fitness}
\alias{viable_fitness}
\title{Compute Region of Positive Fitnes}
\usage{
viable_fitness(bounds, p, x = NULL, log_scale = TRUE, dx = 1)
}
\arguments{
\item{bounds}{Matrix of bounds; two columns corresponding to the
lower and upper limits, each row corresponds to a trait (the name
will be used).}

\item{p}{Parameters object to use.  Importantly, the
\code{strategy_default} element gets used here.}

\item{x}{Initial value - If not given, then the value from the
default strategy within \code{p} is used.}

\item{log_scale}{Is the parameter naturally on a log scale?  If
so, this will greatly speed things up.  Can be a vector of length
`\code{nrow(bounds)}`}

\item{dx}{Amount to step the trait.  If \code{log_scale} is
\code{TRUE}, this is on a log scale.}
}
\description{
Compute region of positive fitness.  This will be the values where
fitness is approximately zero.
}
\author{
Rich FitzJohn
}
